Moved the clean cqc jobs and their validation jobs into new folder structure
