+++ projects/_01_ingest/unittest_data/ingest_test_file_data.py
@@ -4,7 +4,20 @@ from datetime import date
 from utils.column_names.capacity_tracker_columns import (
     CapacityTrackerNonResColumns as CTNR,
 )
-from utils.column_values.categorical_column_values import MainJobRoleLabels
+from utils.column_names.raw_data_files.cqc_location_api_columns import (
+    NewCqcLocationApiColumns as CQCL,
+)
+from utils.column_values.categorical_column_values import (
+    CareHome,
+    LocationType,
+    MainJobRoleLabels,
+    PrimaryServiceType,
+    RegistrationStatus,
+    RelatedLocation,
+    Sector,
+    Services,
+)
+from utils.raw_data_adjustments import RecordsToRemoveInLocationsData
 
 
 @dataclass
@@ -915,3 +928,2384 @@ class PostcodeMatcherData:
     expected_join_postcode_data_unmatched_rows = [
         ("1-002", date(2020, 1, 1), "AA1ZAA"),
     ]
+
+
+@dataclass
+class CQCLocationsData:
+    sample_rows = [
+        (
+            "location1",
+            "provider1",
+            "Location",
+            "Social Care Org",
+            "name of location",
+            "Registered",
+            "2020-01-01",
+            None,
+            "N",
+            20,
+            "www.website.org",
+            "1 The Street",
+            "Leeds",
+            "West Yorkshire",
+            "Yorkshire",
+            "LS1 2AB",
+            "50.123455",
+            "-5.6789",
+            "Y",
+            "Adult social care",
+            "01234567891",
+            "Trafford",
+            [
+                {
+                    "name": "Personal care",
+                    "code": "RA1",
+                    "contacts": [
+                        {
+                            "personfamilyname": "Doe",
+                            "persongivenname": "John",
+                            "personroles": ["Registered Manager"],
+                            "persontitle": "Mr",
+                        }
+                    ],
+                }
+            ],
+            [{"name": "Homecare agencies", "description": "Domiciliary care service"}],
+            [{"name": "Services for everyone"}],
+            {
+                CQCL.overall: {
+                    CQCL.organisation_id: None,
+                    CQCL.rating: "Overall rating Excellent",
+                    CQCL.report_date: "report_date",
+                    CQCL.report_link_id: None,
+                    CQCL.use_of_resources: {
+                        CQCL.organisation_id: None,
+                        CQCL.summary: None,
+                        CQCL.use_of_resources_rating: None,
+                        CQCL.combined_quality_summary: None,
+                        CQCL.combined_quality_rating: None,
+                        CQCL.report_date: None,
+                        CQCL.report_link_id: None,
+                    },
+                    CQCL.key_question_ratings: [
+                        {
+                            CQCL.name: "Safe",
+                            CQCL.rating: "Safe rating Good",
+                            CQCL.report_date: None,
+                            CQCL.organisation_id: None,
+                            CQCL.report_link_id: None,
+                        },
+                        {
+                            CQCL.name: "Well-led",
+                            CQCL.rating: "Well-led rating Good",
+                            CQCL.report_date: None,
+                            CQCL.organisation_id: None,
+                            CQCL.report_link_id: None,
+                        },
+                        {
+                            CQCL.name: "Caring",
+                            CQCL.rating: "Caring rating Good",
+                            CQCL.report_date: None,
+                            CQCL.organisation_id: None,
+                            CQCL.report_link_id: None,
+                        },
+                        {
+                            CQCL.name: "Responsive",
+                            CQCL.rating: "Responsive rating Good",
+                            CQCL.report_date: None,
+                            CQCL.organisation_id: None,
+                            CQCL.report_link_id: None,
+                        },
+                        {
+                            CQCL.name: "Effective",
+                            CQCL.rating: "Effective rating Good",
+                            CQCL.report_date: None,
+                            CQCL.organisation_id: None,
+                            CQCL.report_link_id: None,
+                        },
+                    ],
+                },
+                CQCL.service_ratings: [
+                    {
+                        CQCL.name: None,
+                        CQCL.rating: None,
+                        CQCL.report_date: None,
+                        CQCL.organisation_id: None,
+                        CQCL.report_link_id: None,
+                        CQCL.key_question_ratings: [
+                            {
+                                CQCL.name: None,
+                                CQCL.rating: None,
+                            },
+                        ],
+                    },
+                ],
+            },
+            [
+                {
+                    CQCL.report_date: "report_date",
+                    CQCL.report_link_id: None,
+                    CQCL.organisation_id: None,
+                    CQCL.service_ratings: [
+                        {
+                            CQCL.name: None,
+                            CQCL.rating: None,
+                            CQCL.key_question_ratings: [
+                                {
+                                    CQCL.name: None,
+                                    CQCL.rating: None,
+                                },
+                            ],
+                        },
+                    ],
+                    CQCL.overall: {
+                        CQCL.rating: "Overall rating Excellent",
+                        CQCL.use_of_resources: {
+                            CQCL.combined_quality_rating: None,
+                            CQCL.combined_quality_summary: None,
+                            CQCL.use_of_resources_rating: None,
+                            CQCL.use_of_resources_summary: None,
+                        },
+                        CQCL.key_question_ratings: [
+                            {CQCL.name: "Safe", CQCL.rating: "Safe rating Good"},
+                            {
+                                CQCL.name: "Well-led",
+                                CQCL.rating: "Well-led rating Good",
+                            },
+                            {CQCL.name: "Caring", CQCL.rating: "Caring rating Good"},
+                            {
+                                CQCL.name: "Responsive",
+                                CQCL.rating: "Responsive rating Good",
+                            },
+                            {
+                                CQCL.name: "Effective",
+                                CQCL.rating: "Effective rating Good",
+                            },
+                        ],
+                    },
+                },
+            ],
+            [
+                {
+                    CQCL.related_location_id: "1",
+                    CQCL.related_location_name: "name",
+                    CQCL.type: "type",
+                    CQCL.reason: "reason",
+                }
+            ],
+            "2020-01-01",
+        ),
+    ]
+
+    impute_historic_relationships_rows = [
+        ("1-001", date(2024, 1, 1), RegistrationStatus.registered, None),
+        ("1-002", date(2024, 1, 1), RegistrationStatus.registered, None),
+        (
+            "1-002",
+            date(2024, 2, 1),
+            RegistrationStatus.registered,
+            [
+                {
+                    CQCL.related_location_id: "1-0021",
+                    CQCL.related_location_name: "Name before",
+                    CQCL.type: "HSCA Predecessor",
+                    CQCL.reason: "New Provider",
+                }
+            ],
+        ),
+        ("1-003", date(2024, 1, 1), RegistrationStatus.registered, None),
+        (
+            "1-003",
+            date(2024, 2, 1),
+            RegistrationStatus.deregistered,
+            [
+                {
+                    CQCL.related_location_id: "1-0031",
+                    CQCL.related_location_name: "Name after",
+                    CQCL.type: "HSCA Successor",
+                    CQCL.reason: "Location Move",
+                }
+            ],
+        ),
+        ("1-004", date(2024, 1, 1), RegistrationStatus.registered, None),
+        (
+            "1-004",
+            date(2024, 2, 1),
+            RegistrationStatus.deregistered,
+            [
+                {
+                    CQCL.related_location_id: "1-0041",
+                    CQCL.related_location_name: "Name before",
+                    CQCL.type: "HSCA Predecessor",
+                    CQCL.reason: "New Provider",
+                }
+            ],
+        ),
+        ("1-005", date(2024, 1, 1), RegistrationStatus.registered, None),
+        (
+            "1-005",
+            date(2024, 2, 1),
+            RegistrationStatus.deregistered,
+            [
+                {
+                    CQCL.related_location_id: "1-0051",
+                    CQCL.related_location_name: "Name after",
+                    CQCL.type: "HSCA Successor",
+                    CQCL.reason: "Location Move",
+                },
+                {
+                    CQCL.related_location_id: "1-0052",
+                    CQCL.related_location_name: "Name before",
+                    CQCL.type: "HSCA Predecessor",
+                    CQCL.reason: "New Provider",
+                },
+                {
+                    CQCL.related_location_id: "1-0053",
+                    CQCL.related_location_name: "Name before",
+                    CQCL.type: "HSCA Predecessor",
+                    CQCL.reason: "New Provider",
+                },
+            ],
+        ),
+        ("1-006", date(2024, 1, 1), RegistrationStatus.deregistered, None),
+        (
+            "1-006",
+            date(2024, 2, 1),
+            RegistrationStatus.deregistered,
+            [
+                {
+                    CQCL.related_location_id: "1-0061",
+                    CQCL.related_location_name: "Name after",
+                    CQCL.type: "HSCA Successor",
+                    CQCL.reason: "Location Move",
+                },
+                {
+                    CQCL.related_location_id: "1-0062",
+                    CQCL.related_location_name: "Name before",
+                    CQCL.type: "HSCA Predecessor",
+                    CQCL.reason: "New Provider",
+                },
+                {
+                    CQCL.related_location_id: "1-0063",
+                    CQCL.related_location_name: "Name before",
+                    CQCL.type: "HSCA Predecessor",
+                    CQCL.reason: "New Provider",
+                },
+            ],
+        ),
+        ("1-007", date(2024, 1, 1), RegistrationStatus.registered, None),
+        (
+            "1-007",
+            date(2024, 2, 1),
+            RegistrationStatus.deregistered,
+            [
+                {
+                    CQCL.related_location_id: "1-0071",
+                    CQCL.related_location_name: "Name before",
+                    CQCL.type: "HSCA Predecessor",
+                    CQCL.reason: "New Provider",
+                }
+            ],
+        ),
+        (
+            "1-007",
+            date(2024, 3, 1),
+            RegistrationStatus.deregistered,
+            [
+                {
+                    CQCL.related_location_id: "1-0072",
+                    CQCL.related_location_name: "Name before",
+                    CQCL.type: "HSCA Predecessor",
+                    CQCL.reason: "New Provider",
+                }
+            ],
+        ),
+    ]
+    expected_impute_historic_relationships_rows = [
+        ("1-001", date(2024, 1, 1), RegistrationStatus.registered, None, None),
+        (
+            "1-002",
+            date(2024, 1, 1),
+            RegistrationStatus.registered,
+            None,
+            [
+                {
+                    CQCL.related_location_id: "1-0021",
+                    CQCL.related_location_name: "Name before",
+                    CQCL.type: "HSCA Predecessor",
+                    CQCL.reason: "New Provider",
+                }
+            ],
+        ),
+        (
+            "1-002",
+            date(2024, 2, 1),
+            RegistrationStatus.registered,
+            [
+                {
+                    CQCL.related_location_id: "1-0021",
+                    CQCL.related_location_name: "Name before",
+                    CQCL.type: "HSCA Predecessor",
+                    CQCL.reason: "New Provider",
+                }
+            ],
+            [
+                {
+                    CQCL.related_location_id: "1-0021",
+                    CQCL.related_location_name: "Name before",
+                    CQCL.type: "HSCA Predecessor",
+                    CQCL.reason: "New Provider",
+                }
+            ],
+        ),
+        ("1-003", date(2024, 1, 1), RegistrationStatus.registered, None, None),
+        (
+            "1-003",
+            date(2024, 2, 1),
+            RegistrationStatus.deregistered,
+            [
+                {
+                    CQCL.related_location_id: "1-0031",
+                    CQCL.related_location_name: "Name after",
+                    CQCL.type: "HSCA Successor",
+                    CQCL.reason: "Location Move",
+                }
+            ],
+            [
+                {
+                    CQCL.related_location_id: "1-0031",
+                    CQCL.related_location_name: "Name after",
+                    CQCL.type: "HSCA Successor",
+                    CQCL.reason: "Location Move",
+                }
+            ],
+        ),
+        (
+            "1-004",
+            date(2024, 1, 1),
+            RegistrationStatus.registered,
+            None,
+            [
+                {
+                    CQCL.related_location_id: "1-0041",
+                    CQCL.related_location_name: "Name before",
+                    CQCL.type: "HSCA Predecessor",
+                    CQCL.reason: "New Provider",
+                }
+            ],
+        ),
+        (
+            "1-004",
+            date(2024, 2, 1),
+            RegistrationStatus.deregistered,
+            [
+                {
+                    CQCL.related_location_id: "1-0041",
+                    CQCL.related_location_name: "Name before",
+                    CQCL.type: "HSCA Predecessor",
+                    CQCL.reason: "New Provider",
+                }
+            ],
+            [
+                {
+                    CQCL.related_location_id: "1-0041",
+                    CQCL.related_location_name: "Name before",
+                    CQCL.type: "HSCA Predecessor",
+                    CQCL.reason: "New Provider",
+                }
+            ],
+        ),
+        (
+            "1-005",
+            date(2024, 1, 1),
+            RegistrationStatus.registered,
+            None,
+            [
+                {
+                    CQCL.related_location_id: "1-0053",
+                    CQCL.related_location_name: "Name before",
+                    CQCL.type: "HSCA Predecessor",
+                    CQCL.reason: "New Provider",
+                },
+                {
+                    CQCL.related_location_id: "1-0052",
+                    CQCL.related_location_name: "Name before",
+                    CQCL.type: "HSCA Predecessor",
+                    CQCL.reason: "New Provider",
+                },
+            ],
+        ),
+        (
+            "1-005",
+            date(2024, 2, 1),
+            RegistrationStatus.deregistered,
+            [
+                {
+                    CQCL.related_location_id: "1-0051",
+                    CQCL.related_location_name: "Name after",
+                    CQCL.type: "HSCA Successor",
+                    CQCL.reason: "Location Move",
+                },
+                {
+                    CQCL.related_location_id: "1-0052",
+                    CQCL.related_location_name: "Name before",
+                    CQCL.type: "HSCA Predecessor",
+                    CQCL.reason: "New Provider",
+                },
+                {
+                    CQCL.related_location_id: "1-0053",
+                    CQCL.related_location_name: "Name before",
+                    CQCL.type: "HSCA Predecessor",
+                    CQCL.reason: "New Provider",
+                },
+            ],
+            [
+                {
+                    CQCL.related_location_id: "1-0051",
+                    CQCL.related_location_name: "Name after",
+                    CQCL.type: "HSCA Successor",
+                    CQCL.reason: "Location Move",
+                },
+                {
+                    CQCL.related_location_id: "1-0052",
+                    CQCL.related_location_name: "Name before",
+                    CQCL.type: "HSCA Predecessor",
+                    CQCL.reason: "New Provider",
+                },
+                {
+                    CQCL.related_location_id: "1-0053",
+                    CQCL.related_location_name: "Name before",
+                    CQCL.type: "HSCA Predecessor",
+                    CQCL.reason: "New Provider",
+                },
+            ],
+        ),
+        (
+            "1-006",
+            date(2024, 1, 1),
+            RegistrationStatus.deregistered,
+            None,
+            [
+                {
+                    CQCL.related_location_id: "1-0061",
+                    CQCL.related_location_name: "Name after",
+                    CQCL.type: "HSCA Successor",
+                    CQCL.reason: "Location Move",
+                },
+                {
+                    CQCL.related_location_id: "1-0062",
+                    CQCL.related_location_name: "Name before",
+                    CQCL.type: "HSCA Predecessor",
+                    CQCL.reason: "New Provider",
+                },
+                {
+                    CQCL.related_location_id: "1-0063",
+                    CQCL.related_location_name: "Name before",
+                    CQCL.type: "HSCA Predecessor",
+                    CQCL.reason: "New Provider",
+                },
+            ],
+        ),
+        (
+            "1-006",
+            date(2024, 2, 1),
+            RegistrationStatus.deregistered,
+            [
+                {
+                    CQCL.related_location_id: "1-0061",
+                    CQCL.related_location_name: "Name after",
+                    CQCL.type: "HSCA Successor",
+                    CQCL.reason: "Location Move",
+                },
+                {
+                    CQCL.related_location_id: "1-0062",
+                    CQCL.related_location_name: "Name before",
+                    CQCL.type: "HSCA Predecessor",
+                    CQCL.reason: "New Provider",
+                },
+                {
+                    CQCL.related_location_id: "1-0063",
+                    CQCL.related_location_name: "Name before",
+                    CQCL.type: "HSCA Predecessor",
+                    CQCL.reason: "New Provider",
+                },
+            ],
+            [
+                {
+                    CQCL.related_location_id: "1-0061",
+                    CQCL.related_location_name: "Name after",
+                    CQCL.type: "HSCA Successor",
+                    CQCL.reason: "Location Move",
+                },
+                {
+                    CQCL.related_location_id: "1-0062",
+                    CQCL.related_location_name: "Name before",
+                    CQCL.type: "HSCA Predecessor",
+                    CQCL.reason: "New Provider",
+                },
+                {
+                    CQCL.related_location_id: "1-0063",
+                    CQCL.related_location_name: "Name before",
+                    CQCL.type: "HSCA Predecessor",
+                    CQCL.reason: "New Provider",
+                },
+            ],
+        ),
+        (
+            "1-007",
+            date(2024, 1, 1),
+            RegistrationStatus.registered,
+            None,
+            [
+                {
+                    CQCL.related_location_id: "1-0071",
+                    CQCL.related_location_name: "Name before",
+                    CQCL.type: "HSCA Predecessor",
+                    CQCL.reason: "New Provider",
+                }
+            ],
+        ),
+        (
+            "1-007",
+            date(2024, 2, 1),
+            RegistrationStatus.deregistered,
+            [
+                {
+                    CQCL.related_location_id: "1-0071",
+                    CQCL.related_location_name: "Name before",
+                    CQCL.type: "HSCA Predecessor",
+                    CQCL.reason: "New Provider",
+                }
+            ],
+            [
+                {
+                    CQCL.related_location_id: "1-0071",
+                    CQCL.related_location_name: "Name before",
+                    CQCL.type: "HSCA Predecessor",
+                    CQCL.reason: "New Provider",
+                }
+            ],
+        ),
+        (
+            "1-007",
+            date(2024, 3, 1),
+            RegistrationStatus.deregistered,
+            [
+                {
+                    CQCL.related_location_id: "1-0072",
+                    CQCL.related_location_name: "Name before",
+                    CQCL.type: "HSCA Predecessor",
+                    CQCL.reason: "New Provider",
+                }
+            ],
+            [
+                {
+                    CQCL.related_location_id: "1-0072",
+                    CQCL.related_location_name: "Name before",
+                    CQCL.type: "HSCA Predecessor",
+                    CQCL.reason: "New Provider",
+                }
+            ],
+        ),
+    ]
+    impute_historic_relationships_when_type_is_none_returns_none_rows = [
+        ("1-001", date(2024, 1, 1), RegistrationStatus.registered, None),
+    ]
+    expected_impute_historic_relationships_when_type_is_none_returns_none_rows = [
+        ("1-001", date(2024, 1, 1), RegistrationStatus.registered, None, None),
+    ]
+    impute_historic_relationships_when_type_is_predecessor_returns_predecessor_rows = [
+        ("1-002", date(2024, 1, 1), RegistrationStatus.registered, None),
+        (
+            "1-002",
+            date(2024, 2, 1),
+            RegistrationStatus.registered,
+            [
+                {
+                    CQCL.related_location_id: "1-0021",
+                    CQCL.related_location_name: "Name before",
+                    CQCL.type: "HSCA Predecessor",
+                    CQCL.reason: "New Provider",
+                }
+            ],
+        ),
+    ]
+    expected_impute_historic_relationships_when_type_is_predecessor_returns_predecessor_rows = [
+        (
+            "1-002",
+            date(2024, 1, 1),
+            RegistrationStatus.registered,
+            None,
+            [
+                {
+                    CQCL.related_location_id: "1-0021",
+                    CQCL.related_location_name: "Name before",
+                    CQCL.type: "HSCA Predecessor",
+                    CQCL.reason: "New Provider",
+                }
+            ],
+        ),
+        (
+            "1-002",
+            date(2024, 2, 1),
+            RegistrationStatus.registered,
+            [
+                {
+                    CQCL.related_location_id: "1-0021",
+                    CQCL.related_location_name: "Name before",
+                    CQCL.type: "HSCA Predecessor",
+                    CQCL.reason: "New Provider",
+                }
+            ],
+            [
+                {
+                    CQCL.related_location_id: "1-0021",
+                    CQCL.related_location_name: "Name before",
+                    CQCL.type: "HSCA Predecessor",
+                    CQCL.reason: "New Provider",
+                }
+            ],
+        ),
+    ]
+    impute_historic_relationships_when_type_is_successor_returns_none_when_registered_rows = [
+        ("1-003", date(2024, 1, 1), RegistrationStatus.registered, None),
+        (
+            "1-003",
+            date(2024, 2, 1),
+            RegistrationStatus.deregistered,
+            [
+                {
+                    CQCL.related_location_id: "1-0031",
+                    CQCL.related_location_name: "Name after",
+                    CQCL.type: "HSCA Successor",
+                    CQCL.reason: "Location Move",
+                }
+            ],
+        ),
+    ]
+    expected_impute_historic_relationships_when_type_is_successor_returns_none_when_registered_rows = [
+        ("1-003", date(2024, 1, 1), RegistrationStatus.registered, None, None),
+        (
+            "1-003",
+            date(2024, 2, 1),
+            RegistrationStatus.deregistered,
+            [
+                {
+                    CQCL.related_location_id: "1-0031",
+                    CQCL.related_location_name: "Name after",
+                    CQCL.type: "HSCA Successor",
+                    CQCL.reason: "Location Move",
+                }
+            ],
+            [
+                {
+                    CQCL.related_location_id: "1-0031",
+                    CQCL.related_location_name: "Name after",
+                    CQCL.type: "HSCA Successor",
+                    CQCL.reason: "Location Move",
+                }
+            ],
+        ),
+    ]
+    impute_historic_relationships_when_type_is_successor_returns_successor_when_deregistered_rows = [
+        ("1-004", date(2024, 1, 1), RegistrationStatus.deregistered, None),
+        (
+            "1-004",
+            date(2024, 2, 1),
+            RegistrationStatus.deregistered,
+            [
+                {
+                    CQCL.related_location_id: "1-0041",
+                    CQCL.related_location_name: "Name after",
+                    CQCL.type: "HSCA Successor",
+                    CQCL.reason: "Location Move",
+                }
+            ],
+        ),
+    ]
+    expected_impute_historic_relationships_when_type_is_successor_returns_successor_when_deregistered_rows = [
+        (
+            "1-004",
+            date(2024, 1, 1),
+            RegistrationStatus.deregistered,
+            None,
+            [
+                {
+                    CQCL.related_location_id: "1-0041",
+                    CQCL.related_location_name: "Name after",
+                    CQCL.type: "HSCA Successor",
+                    CQCL.reason: "Location Move",
+                }
+            ],
+        ),
+        (
+            "1-004",
+            date(2024, 2, 1),
+            RegistrationStatus.deregistered,
+            [
+                {
+                    CQCL.related_location_id: "1-0041",
+                    CQCL.related_location_name: "Name after",
+                    CQCL.type: "HSCA Successor",
+                    CQCL.reason: "Location Move",
+                }
+            ],
+            [
+                {
+                    CQCL.related_location_id: "1-0041",
+                    CQCL.related_location_name: "Name after",
+                    CQCL.type: "HSCA Successor",
+                    CQCL.reason: "Location Move",
+                }
+            ],
+        ),
+    ]
+    impute_historic_relationships_when_type_has_both_types_only_returns_predecessors_when_registered_rows = [
+        ("1-005", date(2024, 1, 1), RegistrationStatus.registered, None),
+        (
+            "1-005",
+            date(2024, 2, 1),
+            RegistrationStatus.deregistered,
+            [
+                {
+                    CQCL.related_location_id: "1-0051",
+                    CQCL.related_location_name: "Name after",
+                    CQCL.type: "HSCA Successor",
+                    CQCL.reason: "Location Move",
+                },
+                {
+                    CQCL.related_location_id: "1-0052",
+                    CQCL.related_location_name: "Name before",
+                    CQCL.type: "HSCA Predecessor",
+                    CQCL.reason: "New Provider",
+                },
+                {
+                    CQCL.related_location_id: "1-0053",
+                    CQCL.related_location_name: "Name before",
+                    CQCL.type: "HSCA Predecessor",
+                    CQCL.reason: "New Provider",
+                },
+            ],
+        ),
+    ]
+    expected_impute_historic_relationships_when_type_has_both_types_only_returns_predecessors_when_registered_rows = [
+        (
+            "1-005",
+            date(2024, 1, 1),
+            RegistrationStatus.registered,
+            None,
+            [
+                {
+                    CQCL.related_location_id: "1-0053",
+                    CQCL.related_location_name: "Name before",
+                    CQCL.type: "HSCA Predecessor",
+                    CQCL.reason: "New Provider",
+                },
+                {
+                    CQCL.related_location_id: "1-0052",
+                    CQCL.related_location_name: "Name before",
+                    CQCL.type: "HSCA Predecessor",
+                    CQCL.reason: "New Provider",
+                },
+            ],
+        ),
+        (
+            "1-005",
+            date(2024, 2, 1),
+            RegistrationStatus.deregistered,
+            [
+                {
+                    CQCL.related_location_id: "1-0051",
+                    CQCL.related_location_name: "Name after",
+                    CQCL.type: "HSCA Successor",
+                    CQCL.reason: "Location Move",
+                },
+                {
+                    CQCL.related_location_id: "1-0052",
+                    CQCL.related_location_name: "Name before",
+                    CQCL.type: "HSCA Predecessor",
+                    CQCL.reason: "New Provider",
+                },
+                {
+                    CQCL.related_location_id: "1-0053",
+                    CQCL.related_location_name: "Name before",
+                    CQCL.type: "HSCA Predecessor",
+                    CQCL.reason: "New Provider",
+                },
+            ],
+            [
+                {
+                    CQCL.related_location_id: "1-0051",
+                    CQCL.related_location_name: "Name after",
+                    CQCL.type: "HSCA Successor",
+                    CQCL.reason: "Location Move",
+                },
+                {
+                    CQCL.related_location_id: "1-0052",
+                    CQCL.related_location_name: "Name before",
+                    CQCL.type: "HSCA Predecessor",
+                    CQCL.reason: "New Provider",
+                },
+                {
+                    CQCL.related_location_id: "1-0053",
+                    CQCL.related_location_name: "Name before",
+                    CQCL.type: "HSCA Predecessor",
+                    CQCL.reason: "New Provider",
+                },
+            ],
+        ),
+    ]
+    impute_historic_relationships_when_type_has_both_types_returns_original_values_when_deregistered_rows = [
+        ("1-006", date(2024, 1, 1), RegistrationStatus.deregistered, None),
+        (
+            "1-006",
+            date(2024, 2, 1),
+            RegistrationStatus.deregistered,
+            [
+                {
+                    CQCL.related_location_id: "1-0061",
+                    CQCL.related_location_name: "Name after",
+                    CQCL.type: "HSCA Successor",
+                    CQCL.reason: "Location Move",
+                },
+                {
+                    CQCL.related_location_id: "1-0062",
+                    CQCL.related_location_name: "Name before",
+                    CQCL.type: "HSCA Predecessor",
+                    CQCL.reason: "New Provider",
+                },
+                {
+                    CQCL.related_location_id: "1-0063",
+                    CQCL.related_location_name: "Name before",
+                    CQCL.type: "HSCA Predecessor",
+                    CQCL.reason: "New Provider",
+                },
+            ],
+        ),
+    ]
+    expected_impute_historic_relationships_when_type_has_both_types_returns_original_values_when_deregistered_rows = [
+        (
+            "1-006",
+            date(2024, 1, 1),
+            RegistrationStatus.deregistered,
+            None,
+            [
+                {
+                    CQCL.related_location_id: "1-0061",
+                    CQCL.related_location_name: "Name after",
+                    CQCL.type: "HSCA Successor",
+                    CQCL.reason: "Location Move",
+                },
+                {
+                    CQCL.related_location_id: "1-0062",
+                    CQCL.related_location_name: "Name before",
+                    CQCL.type: "HSCA Predecessor",
+                    CQCL.reason: "New Provider",
+                },
+                {
+                    CQCL.related_location_id: "1-0063",
+                    CQCL.related_location_name: "Name before",
+                    CQCL.type: "HSCA Predecessor",
+                    CQCL.reason: "New Provider",
+                },
+            ],
+        ),
+        (
+            "1-006",
+            date(2024, 2, 1),
+            RegistrationStatus.deregistered,
+            [
+                {
+                    CQCL.related_location_id: "1-0061",
+                    CQCL.related_location_name: "Name after",
+                    CQCL.type: "HSCA Successor",
+                    CQCL.reason: "Location Move",
+                },
+                {
+                    CQCL.related_location_id: "1-0062",
+                    CQCL.related_location_name: "Name before",
+                    CQCL.type: "HSCA Predecessor",
+                    CQCL.reason: "New Provider",
+                },
+                {
+                    CQCL.related_location_id: "1-0063",
+                    CQCL.related_location_name: "Name before",
+                    CQCL.type: "HSCA Predecessor",
+                    CQCL.reason: "New Provider",
+                },
+            ],
+            [
+                {
+                    CQCL.related_location_id: "1-0061",
+                    CQCL.related_location_name: "Name after",
+                    CQCL.type: "HSCA Successor",
+                    CQCL.reason: "Location Move",
+                },
+                {
+                    CQCL.related_location_id: "1-0062",
+                    CQCL.related_location_name: "Name before",
+                    CQCL.type: "HSCA Predecessor",
+                    CQCL.reason: "New Provider",
+                },
+                {
+                    CQCL.related_location_id: "1-0063",
+                    CQCL.related_location_name: "Name before",
+                    CQCL.type: "HSCA Predecessor",
+                    CQCL.reason: "New Provider",
+                },
+            ],
+        ),
+    ]
+    impute_historic_relationships_where_different_relationships_over_time_returns_first_found_rows = [
+        ("1-007", date(2024, 1, 1), RegistrationStatus.registered, None),
+        (
+            "1-007",
+            date(2024, 2, 1),
+            RegistrationStatus.deregistered,
+            [
+                {
+                    CQCL.related_location_id: "1-0071",
+                    CQCL.related_location_name: "Name before",
+                    CQCL.type: "HSCA Predecessor",
+                    CQCL.reason: "New Provider",
+                }
+            ],
+        ),
+        (
+            "1-007",
+            date(2024, 3, 1),
+            RegistrationStatus.deregistered,
+            [
+                {
+                    CQCL.related_location_id: "1-0072",
+                    CQCL.related_location_name: "Name before",
+                    CQCL.type: "HSCA Predecessor",
+                    CQCL.reason: "New Provider",
+                }
+            ],
+        ),
+    ]
+    expected_impute_historic_relationships_where_different_relationships_over_time_returns_first_found_rows = [
+        (
+            "1-007",
+            date(2024, 1, 1),
+            RegistrationStatus.registered,
+            None,
+            [
+                {
+                    CQCL.related_location_id: "1-0071",
+                    CQCL.related_location_name: "Name before",
+                    CQCL.type: "HSCA Predecessor",
+                    CQCL.reason: "New Provider",
+                }
+            ],
+        ),
+        (
+            "1-007",
+            date(2024, 2, 1),
+            RegistrationStatus.deregistered,
+            [
+                {
+                    CQCL.related_location_id: "1-0071",
+                    CQCL.related_location_name: "Name before",
+                    CQCL.type: "HSCA Predecessor",
+                    CQCL.reason: "New Provider",
+                }
+            ],
+            [
+                {
+                    CQCL.related_location_id: "1-0071",
+                    CQCL.related_location_name: "Name before",
+                    CQCL.type: "HSCA Predecessor",
+                    CQCL.reason: "New Provider",
+                }
+            ],
+        ),
+        (
+            "1-007",
+            date(2024, 3, 1),
+            RegistrationStatus.deregistered,
+            [
+                {
+                    CQCL.related_location_id: "1-0072",
+                    CQCL.related_location_name: "Name before",
+                    CQCL.type: "HSCA Predecessor",
+                    CQCL.reason: "New Provider",
+                }
+            ],
+            [
+                {
+                    CQCL.related_location_id: "1-0072",
+                    CQCL.related_location_name: "Name before",
+                    CQCL.type: "HSCA Predecessor",
+                    CQCL.reason: "New Provider",
+                }
+            ],
+        ),
+    ]
+
+    get_relationships_where_type_is_none_returns_none_rows = [
+        ("1-001", date(2024, 1, 1), RegistrationStatus.deregistered, None),
+    ]
+    expected_get_relationships_where_type_is_none_returns_none_rows = [
+        ("1-001", date(2024, 1, 1), RegistrationStatus.deregistered, None, None),
+    ]
+    get_relationships_where_type_is_successor_returns_none_rows = [
+        (
+            "1-002",
+            date(2024, 2, 1),
+            RegistrationStatus.deregistered,
+            [
+                {
+                    CQCL.related_location_id: "1-0021",
+                    CQCL.related_location_name: "Name after",
+                    CQCL.type: "HSCA Successor",
+                    CQCL.reason: "Location Move",
+                }
+            ],
+        ),
+    ]
+    expected_get_relationships_where_type_is_successor_returns_none_rows = [
+        (
+            "1-002",
+            date(2024, 2, 1),
+            RegistrationStatus.deregistered,
+            [
+                {
+                    CQCL.related_location_id: "1-0021",
+                    CQCL.related_location_name: "Name after",
+                    CQCL.type: "HSCA Successor",
+                    CQCL.reason: "Location Move",
+                }
+            ],
+            None,
+        ),
+    ]
+    get_relationships_where_type_is_predecessor_returns_predecessor_rows = [
+        (
+            "1-003",
+            date(2024, 2, 1),
+            RegistrationStatus.deregistered,
+            [
+                {
+                    CQCL.related_location_id: "1-0031",
+                    CQCL.related_location_name: "Name before",
+                    CQCL.type: "HSCA Predecessor",
+                    CQCL.reason: "New Provider",
+                }
+            ],
+        ),
+    ]
+    expected_get_relationships_where_type_is_predecessor_returns_predecessor_rows = [
+        (
+            "1-003",
+            date(2024, 2, 1),
+            RegistrationStatus.deregistered,
+            [
+                {
+                    CQCL.related_location_id: "1-0031",
+                    CQCL.related_location_name: "Name before",
+                    CQCL.type: "HSCA Predecessor",
+                    CQCL.reason: "New Provider",
+                }
+            ],
+            [
+                {
+                    CQCL.related_location_id: "1-0031",
+                    CQCL.related_location_name: "Name before",
+                    CQCL.type: "HSCA Predecessor",
+                    CQCL.reason: "New Provider",
+                }
+            ],
+        ),
+    ]
+    get_relationships_where_type_has_both_types_only_returns_predecessor_rows = [
+        (
+            "1-004",
+            date(2024, 2, 1),
+            RegistrationStatus.deregistered,
+            [
+                {
+                    CQCL.related_location_id: "1-0041",
+                    CQCL.related_location_name: "Name after",
+                    CQCL.type: "HSCA Successor",
+                    CQCL.reason: "Location Move",
+                },
+                {
+                    CQCL.related_location_id: "1-0042",
+                    CQCL.related_location_name: "Name before",
+                    CQCL.type: "HSCA Predecessor",
+                    CQCL.reason: "New Provider",
+                },
+                {
+                    CQCL.related_location_id: "1-0043",
+                    CQCL.related_location_name: "Name before",
+                    CQCL.type: "HSCA Predecessor",
+                    CQCL.reason: "New Provider",
+                },
+            ],
+        ),
+    ]
+    expected_get_relationships_where_type_has_both_types_only_returns_predecessor_rows = [
+        (
+            "1-004",
+            date(2024, 2, 1),
+            RegistrationStatus.deregistered,
+            [
+                {
+                    CQCL.related_location_id: "1-0041",
+                    CQCL.related_location_name: "Name after",
+                    CQCL.type: "HSCA Successor",
+                    CQCL.reason: "Location Move",
+                },
+                {
+                    CQCL.related_location_id: "1-0042",
+                    CQCL.related_location_name: "Name before",
+                    CQCL.type: "HSCA Predecessor",
+                    CQCL.reason: "New Provider",
+                },
+                {
+                    CQCL.related_location_id: "1-0043",
+                    CQCL.related_location_name: "Name before",
+                    CQCL.type: "HSCA Predecessor",
+                    CQCL.reason: "New Provider",
+                },
+            ],
+            [
+                {
+                    CQCL.related_location_id: "1-0042",
+                    CQCL.related_location_name: "Name before",
+                    CQCL.type: "HSCA Predecessor",
+                    CQCL.reason: "New Provider",
+                },
+                {
+                    CQCL.related_location_id: "1-0043",
+                    CQCL.related_location_name: "Name before",
+                    CQCL.type: "HSCA Predecessor",
+                    CQCL.reason: "New Provider",
+                },
+            ],
+        ),
+    ]
+
+    # fmt: off
+    impute_missing_struct_column_rows = [
+        ("1-001", date(2024, 1, 1), []),
+        ("1-001", date(2024, 2, 1), None),
+        ("1-001", date(2024, 3, 1), [{"name": "Name A", "description": "Desc A"}]),
+        ("1-001", date(2024, 4, 1), []),
+        ("1-001", date(2024, 5, 1), None),
+        ("1-001", date(2024, 6, 1), [{"name": "Name B", "description": "Desc B"}, {"name": "Name C", "description": "Desc C"}]),
+        ("1-001", date(2024, 7, 1), None),
+        ("1-001", date(2024, 8, 1), []),
+        ("1-002", date(2024, 1, 1), []),
+        ("1-002", date(2024, 2, 1), None),
+    ]
+    expected_impute_missing_struct_column_rows = [
+        ("1-001", date(2024, 1, 1), [], [{"name": "Name A", "description": "Desc A"}]),
+        ("1-001", date(2024, 2, 1), None, [{"name": "Name A", "description": "Desc A"}]),
+        ("1-001", date(2024, 3, 1), [{"name": "Name A", "description": "Desc A"}], [{"name": "Name A", "description": "Desc A"}]),
+        ("1-001", date(2024, 4, 1), [], [{"name": "Name A", "description": "Desc A"}]),
+        ("1-001", date(2024, 5, 1), None, [{"name": "Name A", "description": "Desc A"}]),
+        ("1-001", date(2024, 6, 1), [{"name": "Name B", "description": "Desc B"}, {"name": "Name C", "description": "Desc C"}], [{"name": "Name B", "description": "Desc B"}, {"name": "Name C", "description": "Desc C"}]),
+        ("1-001", date(2024, 7, 1), None, [{"name": "Name B", "description": "Desc B"}, {"name": "Name C", "description": "Desc C"}]),
+        ("1-001", date(2024, 8, 1), [], [{"name": "Name B", "description": "Desc B"}, {"name": "Name C", "description": "Desc C"}]),
+        ("1-002", date(2024, 1, 1), [], None),
+        ("1-002", date(2024, 2, 1), None, None),
+    ]
+    # fmt: on
+
+    remove_locations_that_never_had_regulated_activities_rows = [
+        (
+            "loc 1",
+            [
+                {
+                    "name": "Personal care",
+                    "code": "RA1",
+                    "contacts": [
+                        {
+                            "personfamilyname": "Doe",
+                            "persongivenname": "John",
+                            "personroles": ["Registered Manager"],
+                            "persontitle": "Mr",
+                        }
+                    ],
+                }
+            ],
+        ),
+        ("loc 2", None),
+    ]
+    expected_remove_locations_that_never_had_regulated_activities_rows = [
+        (
+            "loc 1",
+            [
+                {
+                    "name": "Personal care",
+                    "code": "RA1",
+                    "contacts": [
+                        {
+                            "personfamilyname": "Doe",
+                            "persongivenname": "John",
+                            "personroles": ["Registered Manager"],
+                            "persontitle": "Mr",
+                        }
+                    ],
+                }
+            ],
+        ),
+    ]
+
+    extract_from_struct_rows = [
+        (
+            "1-001",
+            [
+                {
+                    CQCL.name: "Homecare agencies",
+                    CQCL.description: Services.domiciliary_care_service,
+                }
+            ],
+        ),
+        (
+            "1-002",
+            [
+                {
+                    CQCL.name: "With nursing",
+                    CQCL.description: Services.care_home_service_with_nursing,
+                },
+                {
+                    CQCL.name: "Without nursing",
+                    CQCL.description: Services.care_home_service_without_nursing,
+                },
+            ],
+        ),
+        (
+            "1-003",
+            None,
+        ),
+    ]
+    expected_extract_from_struct_rows = [
+        (
+            "1-001",
+            [
+                {
+                    CQCL.name: "Homecare agencies",
+                    CQCL.description: Services.domiciliary_care_service,
+                },
+            ],
+            [Services.domiciliary_care_service],
+        ),
+        (
+            "1-002",
+            [
+                {
+                    CQCL.name: "With nursing",
+                    CQCL.description: Services.care_home_service_with_nursing,
+                },
+                {
+                    CQCL.name: "Without nursing",
+                    CQCL.description: Services.care_home_service_without_nursing,
+                },
+            ],
+            [
+                Services.care_home_service_with_nursing,
+                Services.care_home_service_without_nursing,
+            ],
+        ),
+        (
+            "1-003",
+            None,
+            None,
+        ),
+    ]
+
+    primary_service_type_rows = [
+        (
+            "location1",
+            "provider1",
+            [
+                {
+                    "name": "Homecare agencies",
+                    "description": "Domiciliary care service",
+                }
+            ],
+        ),
+        (
+            "location2",
+            "provider2",
+            [
+                {
+                    "name": "With nursing",
+                    "description": "Care home service with nursing",
+                }
+            ],
+        ),
+        (
+            "location3",
+            "provider3",
+            [
+                {
+                    "name": "Without nursing",
+                    "description": "Care home service without nursing",
+                }
+            ],
+        ),
+        (
+            "location4",
+            "provider4",
+            [
+                {
+                    "name": "With nursing",
+                    "description": "Care home service with nursing",
+                },
+                {
+                    "name": "Without nursing",
+                    "description": "Care home service without nursing",
+                },
+            ],
+        ),
+        (
+            "location5",
+            "provider5",
+            [
+                {
+                    "name": "Without nursing",
+                    "description": "Care home service without nursing",
+                },
+                {
+                    "name": "Fake",
+                    "description": "Fake service",
+                },
+            ],
+        ),
+        (
+            "location6",
+            "provider6",
+            [
+                {
+                    "name": "With nursing",
+                    "description": "Care home service with nursing",
+                },
+                {
+                    "name": "Homecare agencies",
+                    "description": "Domiciliary care service",
+                },
+            ],
+        ),
+        (
+            "location7",
+            "provider7",
+            [
+                {
+                    "name": "Without nursing",
+                    "description": "Care home service without nursing",
+                },
+                {
+                    "name": "With nursing",
+                    "description": "Care home service with nursing",
+                },
+            ],
+        ),
+        (
+            "location8",
+            "provider8",
+            [
+                {
+                    "name": "Without nursing",
+                    "description": "Care home service without nursing",
+                },
+                {
+                    "name": "Homecare agencies",
+                    "description": "Domiciliary care service",
+                },
+            ],
+        ),
+        (
+            "location9",
+            "provider9",
+            [
+                {
+                    "name": "Homecare agencies",
+                    "description": "Domiciliary care service",
+                },
+                {
+                    "name": "Without nursing",
+                    "description": "Care home service without nursing",
+                },
+            ],
+        ),
+        (
+            "location10",
+            "provider10",
+            [
+                {
+                    "name": "Homecare agencies",
+                    "description": "Domiciliary care service",
+                },
+                {
+                    "name": "With nursing",
+                    "description": "Care home service with nursing",
+                },
+            ],
+        ),
+    ]
+    expected_primary_service_type_rows = [
+        (
+            "location1",
+            "provider1",
+            [
+                {
+                    "name": "Homecare agencies",
+                    "description": "Domiciliary care service",
+                }
+            ],
+            PrimaryServiceType.non_residential,
+        ),
+        (
+            "location2",
+            "provider2",
+            [
+                {
+                    "name": "With nursing",
+                    "description": "Care home service with nursing",
+                }
+            ],
+            PrimaryServiceType.care_home_with_nursing,
+        ),
+        (
+            "location3",
+            "provider3",
+            [
+                {
+                    "name": "Without nursing",
+                    "description": "Care home service without nursing",
+                }
+            ],
+            PrimaryServiceType.care_home_only,
+        ),
+        (
+            "location4",
+            "provider4",
+            [
+                {
+                    "name": "With nursing",
+                    "description": "Care home service with nursing",
+                },
+                {
+                    "name": "Without nursing",
+                    "description": "Care home service without nursing",
+                },
+            ],
+            PrimaryServiceType.care_home_with_nursing,
+        ),
+        (
+            "location5",
+            "provider5",
+            [
+                {
+                    "name": "Without nursing",
+                    "description": "Care home service without nursing",
+                },
+                {
+                    "name": "Fake",
+                    "description": "Fake service",
+                },
+            ],
+            PrimaryServiceType.care_home_only,
+        ),
+        (
+            "location6",
+            "provider6",
+            [
+                {
+                    "name": "With nursing",
+                    "description": "Care home service with nursing",
+                },
+                {
+                    "name": "Homecare agencies",
+                    "description": "Domiciliary care service",
+                },
+            ],
+            PrimaryServiceType.care_home_with_nursing,
+        ),
+        (
+            "location7",
+            "provider7",
+            [
+                {
+                    "name": "Without nursing",
+                    "description": "Care home service without nursing",
+                },
+                {
+                    "name": "With nursing",
+                    "description": "Care home service with nursing",
+                },
+            ],
+            PrimaryServiceType.care_home_with_nursing,
+        ),
+        (
+            "location8",
+            "provider8",
+            [
+                {
+                    "name": "Without nursing",
+                    "description": "Care home service without nursing",
+                },
+                {
+                    "name": "Homecare agencies",
+                    "description": "Domiciliary care service",
+                },
+            ],
+            PrimaryServiceType.care_home_only,
+        ),
+        (
+            "location9",
+            "provider9",
+            [
+                {
+                    "name": "Homecare agencies",
+                    "description": "Domiciliary care service",
+                },
+                {
+                    "name": "Without nursing",
+                    "description": "Care home service without nursing",
+                },
+            ],
+            PrimaryServiceType.care_home_only,
+        ),
+        (
+            "location10",
+            "provider10",
+            [
+                {
+                    "name": "Homecare agencies",
+                    "description": "Domiciliary care service",
+                },
+                {
+                    "name": "With nursing",
+                    "description": "Care home service with nursing",
+                },
+            ],
+            PrimaryServiceType.care_home_with_nursing,
+        ),
+    ]
+
+    realign_carehome_column_rows = [
+        ("1", CareHome.care_home, PrimaryServiceType.care_home_only),
+        ("2", CareHome.not_care_home, PrimaryServiceType.care_home_only),
+        ("3", CareHome.care_home, PrimaryServiceType.care_home_with_nursing),
+        ("4", CareHome.not_care_home, PrimaryServiceType.care_home_with_nursing),
+        ("5", CareHome.care_home, PrimaryServiceType.non_residential),
+        ("6", CareHome.not_care_home, PrimaryServiceType.non_residential),
+    ]
+    expected_realign_carehome_column_rows = [
+        ("1", CareHome.care_home, PrimaryServiceType.care_home_only),
+        ("2", CareHome.care_home, PrimaryServiceType.care_home_only),
+        ("3", CareHome.care_home, PrimaryServiceType.care_home_with_nursing),
+        ("4", CareHome.care_home, PrimaryServiceType.care_home_with_nursing),
+        ("5", CareHome.not_care_home, PrimaryServiceType.non_residential),
+        ("6", CareHome.not_care_home, PrimaryServiceType.non_residential),
+    ]
+
+    small_location_rows = [
+        (
+            "loc-1",
+            "prov-1",
+            "20200101",
+        ),
+        (
+            "loc-2",
+            "prov-1",
+            "20200101",
+        ),
+        (
+            "loc-3",
+            "prov-2",
+            "20200101",
+        ),
+        (
+            "loc-4",
+            "prov-2",
+            "20210101",
+        ),
+    ]
+
+    join_provider_rows = [
+        (
+            "prov-1",
+            "Apple Tree Care Homes",
+            "Local authority",
+            date(2020, 1, 1),
+        ),
+        (
+            "prov-2",
+            "Sunshine Domestic Care",
+            "Independent",
+            date(2020, 1, 1),
+        ),
+        (
+            "prov-3",
+            "Sunny Days Domestic Care",
+            "Independent",
+            date(2020, 1, 1),
+        ),
+    ]
+
+    expected_joined_rows = [
+        (
+            "loc-1",
+            "prov-1",
+            "Apple Tree Care Homes",
+            "Local authority",
+            date(2020, 1, 1),
+            date(2020, 1, 1),
+        ),
+        (
+            "loc-2",
+            "prov-1",
+            "Apple Tree Care Homes",
+            "Local authority",
+            date(2020, 1, 1),
+            date(2020, 1, 1),
+        ),
+        (
+            "loc-3",
+            "prov-2",
+            "Sunshine Domestic Care",
+            "Independent",
+            date(2020, 1, 1),
+            date(2020, 1, 1),
+        ),
+        (
+            "loc-4",
+            "prov-2",
+            "Sunshine Domestic Care",
+            "Independent",
+            date(2021, 1, 1),
+            date(2020, 1, 1),
+        ),
+    ]
+
+    test_invalid_postcode_data = [
+        ("loc-1", "B69 E3G"),
+        ("loc-2", "UB4 0EJ."),
+        ("loc-3", "PO20 3BD"),
+        ("loc-4", "PR! 9HL"),
+        ("loc-5", None),
+    ]
+
+    expected_invalid_postcode_data = [
+        ("loc-1", "B69 3EG"),
+        ("loc-2", "UB4 0EJ"),
+        ("loc-3", "PO20 3BD"),
+        ("loc-4", "PR1 9HL"),
+        ("loc-5", None),
+    ]
+
+    registration_status_with_missing_data_rows = [
+        (
+            "loc-1",
+            "Registered",
+        ),
+        (
+            "loc-2",
+            "Deregistered",
+        ),
+        (
+            "loc-3",
+            "new value",
+        ),
+    ]
+
+    registration_status_rows = [
+        (
+            "loc-1",
+            "Registered",
+        ),
+        (
+            "loc-2",
+            "Deregistered",
+        ),
+    ]
+
+    expected_registered_rows = [
+        (
+            "loc-1",
+            "Registered",
+        ),
+    ]
+
+    social_care_org_rows = [
+        (
+            "loc-1",
+            "Any none ASC org",
+        ),
+        (
+            "loc-2",
+            "Social Care Org",
+        ),
+        (
+            "loc-3",
+            None,
+        ),
+    ]
+
+    expected_social_care_org_rows = [
+        (
+            "loc-2",
+            "Social Care Org",
+        ),
+    ]
+
+    ons_postcode_directory_rows = [
+        (
+            "LS12AB",
+            date(2021, 1, 1),
+            "Leeds",
+            "Yorkshire & Humber",
+            date(2021, 1, 1),
+            "Leeds",
+            "Yorkshire & Humber",
+        ),
+        (
+            "B693EG",
+            date(2021, 1, 1),
+            "York",
+            "Yorkshire & Humber",
+            date(2021, 1, 1),
+            "York",
+            "Yorkshire & Humber",
+        ),
+        (
+            "PR19HL",
+            date(2019, 1, 1),
+            "Hull",
+            "Yorkshire & Humber",
+            date(2021, 1, 1),
+            "East Riding of Yorkshire",
+            "Yorkshire & Humber",
+        ),
+    ]
+
+    locations_for_ons_join_rows = [
+        ("loc-1", "prov-1", date(2020, 1, 1), "PR1 9AB", "Registered"),
+        ("loc-2", "prov-1", date(2018, 1, 1), "B69 3EG", "Deregistered"),
+        (
+            "loc-3",
+            "prov-2",
+            date(2020, 1, 1),
+            "PR1 9HL",
+            "Deregistered",
+        ),
+        ("loc-4", "prov-2", date(2021, 1, 1), "LS1 2AB", "Registered"),
+    ]
+
+    expected_ons_join_with_null_rows = [
+        (
+            date(2019, 1, 1),
+            "PR19AB",
+            date(2020, 1, 1),
+            "loc-1",
+            "prov-1",
+            None,
+            None,
+            None,
+            None,
+            None,
+            "Registered",
+        ),
+        (
+            None,
+            "B693EG",
+            date(2018, 1, 1),
+            "loc-2",
+            "prov-1",
+            None,
+            None,
+            None,
+            None,
+            None,
+            "Deregistered",
+        ),
+        (
+            date(2019, 1, 1),
+            "PR19HL",
+            date(2020, 1, 1),
+            "loc-3",
+            "prov-2",
+            "Hull",
+            "Yorkshire & Humber",
+            date(2021, 1, 1),
+            "East Riding of Yorkshire",
+            "Yorkshire & Humber",
+            "Deregistered",
+        ),
+        (
+            date(2021, 1, 1),
+            "LS12AB",
+            date(2021, 1, 1),
+            "loc-4",
+            "prov-2",
+            "Leeds",
+            "Yorkshire & Humber",
+            date(2021, 1, 1),
+            "Leeds",
+            "Yorkshire & Humber",
+            "Registered",
+        ),
+    ]
+
+    expected_split_registered_no_nulls_rows = [
+        (
+            date(2019, 1, 1),
+            "PR19AB",
+            date(2020, 1, 1),
+            "loc-1",
+            "prov-1",
+            "Somerset",
+            "Oxen Lane",
+            date(2021, 1, 1),
+            "Somerset",
+            "English Region",
+            "Registered",
+        ),
+        (
+            date(2021, 1, 1),
+            "LS12AB",
+            date(2021, 1, 1),
+            "loc-4",
+            "prov-2",
+            "Leeds",
+            "Yorkshire & Humber",
+            date(2021, 1, 1),
+            "Leeds",
+            "Yorkshire & Humber",
+            "Registered",
+        ),
+    ]
+
+    # fmt: off
+    remove_time_from_date_column_rows = [
+        ("loc_1", "2018-01-01", "20240101", "2018-01-01"),
+        ("loc_1", "2018-01-01 00:00:00", "20231201", "2018-01-01 00:00:00"),
+        ("loc_1", None, "20231101", None),
+    ]
+    expected_remove_time_from_date_column_rows = [
+        ("loc_1", "2018-01-01", "20240101", "2018-01-01"),
+        ("loc_1", "2018-01-01 00:00:00", "20231201", "2018-01-01"),
+        ("loc_1", None, "20231101", None),
+    ]
+    remove_late_registration_dates_rows = [
+        ("loc_1", "20240101", "2024-01-01"),
+        ("loc_2", "20240101", "2024-01-02"),
+        ("loc_3", "20240101", "2023-12-31"),
+        ("loc_4", "20240201", "2024-02-02"),
+        ("loc_4", "20240301", "2024-02-02"),
+    ]
+    expected_remove_late_registration_dates_rows = [
+        ("loc_1", "20240101", "2024-01-01"),
+        ("loc_2", "20240101", None),
+        ("loc_3", "20240101", "2023-12-31"),
+        ("loc_4", "20240201", None),
+        ("loc_4", "20240301", None),
+    ]
+    clean_registration_date_column_rows = [
+        ("loc_1", "2018-01-01", "20240101"),
+        ("loc_1", "2018-01-01 00:00:00", "20231201"),
+        ("loc_1", None, "20231101"),
+        ("loc_2", None, "20240101"),
+        ("loc_2", None, "20231201"),
+        ("loc_2", None, "20231101"),
+    ]
+    expected_clean_registration_date_column_rows = [
+        ("loc_1", "2018-01-01", "20240101", "2018-01-01"),
+        ("loc_1", "2018-01-01 00:00:00", "20231201", "2018-01-01"),
+        ("loc_1", None, "20231101", "2018-01-01"),
+        ("loc_2", None, "20240101", "2023-11-01"),
+        ("loc_2", None, "20231201", "2023-11-01"),
+        ("loc_2", None, "20231101", "2023-11-01"),
+    ]
+    impute_missing_registration_dates_rows=expected_remove_time_from_date_column_rows
+    expected_impute_missing_registration_dates_rows=[
+        ("loc_1", "2018-01-01", "20240101", "2018-01-01"),
+        ("loc_1", "2018-01-01 00:00:00", "20231201", "2018-01-01"),
+        ("loc_1", None, "20231101", "2018-01-01"),
+    ]
+    impute_missing_registration_dates_different_rows=[
+        ("loc_1", "2018-01-01", "20240101", "2018-01-01"),
+        ("loc_1", "2017-01-01 00:00:00", "20231201", "2017-01-01"),
+        ("loc_1", None, "20231101", None),
+    ]
+    expected_impute_missing_registration_dates_different_rows=[
+        ("loc_1", "2018-01-01", "20240101", "2018-01-01"),
+        ("loc_1", "2017-01-01 00:00:00", "20231201", "2017-01-01"),
+        ("loc_1", None, "20231101", "2017-01-01"),
+    ]
+    impute_missing_registration_dates_missing_rows=[
+        ("loc_2", None, "20240101", None),
+        ("loc_2", None, "20231201", None),
+        ("loc_2", None, "20231101", None),
+    ]
+    expected_impute_missing_registration_dates_missing_rows=[
+        ("loc_2", None, "20240101", "2023-11-01"),
+        ("loc_2", None, "20231201", "2023-11-01"),
+        ("loc_2", None, "20231101", "2023-11-01"),
+    ]
+    # fmt: on
+
+    calculate_time_registered_same_day_rows = [
+        ("1-0001", date(2025, 1, 1), date(2025, 1, 1)),
+    ]
+    expected_calculate_time_registered_same_day_rows = [
+        ("1-0001", date(2025, 1, 1), date(2025, 1, 1), 1),
+    ]
+
+    calculate_time_registered_exact_months_apart_rows = [
+        ("1-0001", date(2024, 2, 1), date(2024, 1, 1)),
+        ("1-0002", date(2020, 1, 1), date(2019, 1, 1)),
+    ]
+    expected_calculate_time_registered_exact_months_apart_rows = [
+        ("1-0001", date(2024, 2, 1), date(2024, 1, 1), 2),
+        ("1-0002", date(2020, 1, 1), date(2019, 1, 1), 13),
+    ]
+
+    calculate_time_registered_one_day_less_than_a_full_month_apart_rows = [
+        ("1-0001", date(2025, 1, 1), date(2024, 12, 2)),
+        ("1-0002", date(2025, 6, 8), date(2025, 1, 9)),
+    ]
+    expected_calculate_time_registered_one_day_less_than_a_full_month_apart_rows = [
+        ("1-0001", date(2025, 1, 1), date(2024, 12, 2), 1),
+        ("1-0002", date(2025, 6, 8), date(2025, 1, 9), 5),
+    ]
+
+    calculate_time_registered_one_day_more_than_a_full_month_apart_rows = [
+        ("1-0001", date(2025, 1, 2), date(2024, 12, 1)),
+        ("1-0002", date(2025, 6, 1), date(2025, 1, 31)),
+    ]
+    expected_calculate_time_registered_one_day_more_than_a_full_month_apart_rows = [
+        ("1-0001", date(2025, 1, 2), date(2024, 12, 1), 2),
+        ("1-0002", date(2025, 6, 1), date(2025, 1, 31), 5),
+    ]
+
+    clean_provider_id_column_rows = [
+        ("loc_1", None, "20240101"),
+        ("loc_1", "123456789", "20240201"),
+        ("loc_1", None, "20240201"),
+        ("loc_2", "223456789 223456789", "20240101"),
+        ("loc_2", "223456789", "20240201"),
+        ("loc_2", None, "20240301"),
+    ]
+    expected_clean_provider_id_column_rows = [
+        ("loc_1", "123456789", "20240101"),
+        ("loc_1", "123456789", "20240201"),
+        ("loc_1", "123456789", "20240201"),
+        ("loc_2", "223456789", "20240101"),
+        ("loc_2", "223456789", "20240201"),
+        ("loc_2", "223456789", "20240301"),
+    ]
+    long_provider_id_column_rows = [
+        ("loc_2", "223456789 223456789", "20240101"),
+        ("loc_2", "223456789", "20240201"),
+        ("loc_2", None, "20240301"),
+    ]
+    expected_long_provider_id_column_rows = [
+        ("loc_2", None, "20240101"),
+        ("loc_2", "223456789", "20240201"),
+        ("loc_2", None, "20240301"),
+    ]
+    fill_missing_provider_id_column_rows = [
+        ("loc_1", None, "20240101"),
+        ("loc_1", "123456789", "20240201"),
+        ("loc_1", None, "20240201"),
+    ]
+
+    expected_fill_missing_provider_id_column_rows = [
+        ("loc_1", "123456789", "20240101"),
+        ("loc_1", "123456789", "20240201"),
+        ("loc_1", "123456789", "20240201"),
+    ]
+
+    impute_missing_data_from_provider_dataset_single_value_rows = [
+        (
+            "prov_1",
+            None,
+            date(2024, 1, 1),
+        ),
+        (
+            "prov_1",
+            Sector.independent,
+            date(2024, 2, 1),
+        ),
+        (
+            "prov_1",
+            Sector.independent,
+            date(2024, 3, 1),
+        ),
+    ]
+
+    expected_impute_missing_data_from_provider_dataset_rows = [
+        (
+            "prov_1",
+            Sector.independent,
+            date(2024, 1, 1),
+        ),
+        (
+            "prov_1",
+            Sector.independent,
+            date(2024, 2, 1),
+        ),
+        (
+            "prov_1",
+            Sector.independent,
+            date(2024, 3, 1),
+        ),
+    ]
+
+    impute_missing_data_from_provider_dataset_multiple_values_rows = [
+        (
+            "prov_1",
+            None,
+            date(2024, 1, 1),
+        ),
+        (
+            "prov_1",
+            Sector.independent,
+            date(2024, 3, 1),
+        ),
+        (
+            "prov_1",
+            Sector.local_authority,
+            date(2024, 2, 1),
+        ),
+    ]
+
+    expected_impute_missing_data_from_provider_dataset_multiple_values_rows = [
+        (
+            "prov_1",
+            Sector.local_authority,
+            date(2024, 1, 1),
+        ),
+        (
+            "prov_1",
+            Sector.local_authority,
+            date(2024, 2, 1),
+        ),
+        (
+            "prov_1",
+            Sector.independent,
+            date(2024, 3, 1),
+        ),
+    ]
+    test_only_service_specialist_colleges_rows = [
+        (
+            "loc 1",
+            [Services.specialist_college_service],
+        ),
+        (
+            "loc 4",
+            [Services.care_home_service_with_nursing],
+        ),
+    ]
+    test_multiple_services_specialist_colleges_rows = [
+        (
+            "loc 2",
+            [
+                Services.specialist_college_service,
+                Services.acute_services_with_overnight_beds,
+            ],
+        ),
+        (
+            "loc 3",
+            [
+                Services.acute_services_with_overnight_beds,
+                Services.specialist_college_service,
+            ],
+        ),
+    ]
+    test_without_specialist_colleges_rows = [
+        (
+            "loc 4",
+            [Services.care_home_service_with_nursing],
+        ),
+    ]
+    test_empty_array_specialist_colleges_rows = [
+        (
+            "loc 5",
+            [],
+        ),
+    ]
+    test_null_row_specialist_colleges_rows = [
+        (
+            "loc 6",
+            None,
+        ),
+    ]
+    expected_only_service_specialist_colleges_rows = [
+        (
+            "loc 4",
+            [Services.care_home_service_with_nursing],
+        ),
+    ]
+    expected_multiple_services_specialist_colleges_rows = (
+        test_multiple_services_specialist_colleges_rows
+    )
+    expected_without_specialist_colleges_rows = test_without_specialist_colleges_rows
+    expected_empty_array_specialist_colleges_rows = (
+        test_empty_array_specialist_colleges_rows
+    )
+    expected_null_row_specialist_colleges_rows = test_null_row_specialist_colleges_rows
+
+    add_related_location_column_rows = [
+        ("loc 1", None),
+        ("loc 2", []),
+        (
+            "loc 3",
+            [
+                {
+                    CQCL.related_location_id: "1",
+                    CQCL.related_location_name: "name",
+                    CQCL.type: "type",
+                    CQCL.reason: "reason",
+                }
+            ],
+        ),
+        (
+            "loc 4",
+            [
+                {
+                    CQCL.related_location_id: "1",
+                    CQCL.related_location_name: "name",
+                    CQCL.type: "type",
+                    CQCL.reason: "reason",
+                },
+                {
+                    CQCL.related_location_id: "2",
+                    CQCL.related_location_name: "name",
+                    CQCL.type: "type",
+                    CQCL.reason: "reason",
+                },
+            ],
+        ),
+    ]
+    expected_add_related_location_column_rows = [
+        ("loc 1", None, RelatedLocation.no_related_location),
+        ("loc 2", [], RelatedLocation.no_related_location),
+        (
+            "loc 3",
+            [
+                {
+                    CQCL.related_location_id: "1",
+                    CQCL.related_location_name: "name",
+                    CQCL.type: "type",
+                    CQCL.reason: "reason",
+                }
+            ],
+            RelatedLocation.has_related_location,
+        ),
+        (
+            "loc 4",
+            [
+                {
+                    CQCL.related_location_id: "1",
+                    CQCL.related_location_name: "name",
+                    CQCL.type: "type",
+                    CQCL.reason: "reason",
+                },
+                {
+                    CQCL.related_location_id: "2",
+                    CQCL.related_location_name: "name",
+                    CQCL.type: "type",
+                    CQCL.reason: "reason",
+                },
+            ],
+            RelatedLocation.has_related_location,
+        ),
+    ]
+
+
+@dataclass
+class CQCProviderData:
+    sample_rows_full = [
+        (
+            "1-10000000001",
+            ["1-12000000001"],
+            "Provider",
+            "Organisation",
+            "Independent Healthcare Org",
+            "10000000001",
+            "Care Solutions Direct Limited",
+            "Registered",
+            "2022-01-14",
+            None,
+            "Threefield House",
+            "Southampton",
+            None,
+            "South East",
+            "AA10 3LP",
+            50.93761444091797,
+            -1.452439546585083,
+            "0238206106",
+            "10000001",
+            "Adult social care",
+            "Southampton, Itchen",
+            "Southampton",
+            "20230405",
+        ),
+        (
+            "1-10000000002",
+            ["1-12000000002"],
+            "Provider",
+            "Partnership",
+            "Social Care Org",
+            "10000000002",
+            "Care Solutions Direct Limited",
+            "Registered",
+            "2022-01-14",
+            None,
+            "Threefield House",
+            "Southampton",
+            "Some County",
+            "South East",
+            "AA10 3LP",
+            50.93761444091797,
+            -1.452439546585083,
+            "0238206106",
+            "10000002",
+            "Adult social care",
+            "Southampton, Itchen",
+            "Southampton",
+            "20230405",
+        ),
+        (
+            "1-10000000003",
+            ["1-12000000003"],
+            "Provider",
+            "Individual",
+            "Social Care Org",
+            "10000000003",
+            "Care Solutions Direct Limited",
+            "Deregistered",
+            "2022-01-14",
+            "2022-03-07",
+            "Threefield House",
+            "Southampton",
+            None,
+            "South East",
+            "SO14 3LP",
+            50.93761444091797,
+            -1.452439546585083,
+            "0238206106",
+            "10000003",
+            "Adult social care",
+            "Southampton, Itchen",
+            "Southampton",
+            "20230405",
+        ),
+    ]
+
+    sector_rows = [
+        "1-10000000002",
+        "1-10000000003",
+        "1-10000000004",
+        "1-10000000005",
+    ]
+
+    rows_without_cqc_sector = [
+        ("1-10000000001", "data"),
+        ("1-10000000002", None),
+        ("1-10000000003", "data"),
+    ]
+
+    expected_rows_with_cqc_sector = [
+        ("1-10000000001", "data", Sector.independent),
+        ("1-10000000002", None, Sector.local_authority),
+        ("1-10000000003", "data", Sector.local_authority),
+    ]
+
+
+@dataclass
+class ValidateLocationsAPICleanedData:
+    # fmt: off
+    raw_cqc_locations_rows = [
+        ("1-000000001", "1-001", "20240101", LocationType.social_care_identifier, RegistrationStatus.registered, [{CQCL.name: "name", CQCL.description: Services.care_home_service_with_nursing}], [{CQCL.name: "name", CQCL.code: "A1", CQCL.contacts: []}]),
+        ("1-000000002", "1-001", "20240101", LocationType.social_care_identifier, RegistrationStatus.deregistered, [{CQCL.name: "name", CQCL.description: Services.care_home_service_with_nursing}], [{CQCL.name: "name", CQCL.code: "A1", CQCL.contacts: []}]),
+        ("1-000000001", "1-001", "20240201", LocationType.social_care_identifier, RegistrationStatus.registered, [{CQCL.name: "name", CQCL.description: Services.care_home_service_with_nursing}], [{CQCL.name: "name", CQCL.code: "A1", CQCL.contacts: []}]),
+        ("1-000000002", "1-001", "20240201", "not social care org", RegistrationStatus.deregistered, [{CQCL.name: "name", CQCL.description: Services.care_home_service_with_nursing}], [{CQCL.name: "name", CQCL.code: "A1", CQCL.contacts: []}]),
+    ]
+    # fmt: on
+
+    # fmt: off
+    cleaned_cqc_locations_rows = [
+        ("1-000000002", date(2024, 1, 1), date(2024, 1, 1), "Y", "name", "prov_1", "prov_name", Sector.independent, RegistrationStatus.registered, date(2024, 1, 1), "Y", 5, PrimaryServiceType.care_home_only, date(2024, 1, 1), "cssr", "region", date(2024, 1, 1), "cssr", "region", "RUI", None, [{CQCL.name: "name", CQCL.code: "A1", CQCL.contacts: []}]),
+        ("1-000000001", date(2024, 1, 9), date(2024, 1, 1), "Y", "name", "prov_1", "prov_name", Sector.independent, RegistrationStatus.registered, date(2024, 1, 1), "Y", 5, PrimaryServiceType.care_home_only, date(2024, 1, 1), "cssr", "region", date(2024, 1, 1), "cssr", "region", "RUI", None, [{CQCL.name: "name", CQCL.code: "A1", CQCL.contacts: []}]),
+        ("1-000000001", date(2024, 1, 1), date(2024, 1, 1), "Y", "name", "prov_1", "prov_name", Sector.independent, RegistrationStatus.registered, date(2024, 1, 1), "Y", 5, PrimaryServiceType.care_home_only, date(2024, 1, 1), "cssr", "region", date(2024, 1, 1), "cssr", "region", "RUI", None, [{CQCL.name: "name", CQCL.code: "A1", CQCL.contacts: []}]),
+        ("1-000000002", date(2024, 1, 9), date(2024, 1, 1), "Y", "name", "prov_1", "prov_name", Sector.independent, RegistrationStatus.registered, date(2024, 1, 1), "Y", 5, PrimaryServiceType.care_home_only, date(2024, 1, 1), "cssr", "region", date(2024, 1, 1), "cssr", "region", "RUI", None, [{CQCL.name: "name", CQCL.code: "A1", CQCL.contacts: []}]),
+    ]
+    # fmt: on
+
+
+    # fmt: off
+    calculate_expected_size_rows = [
+        ("loc_1", "prov_1", LocationType.social_care_identifier, RegistrationStatus.registered, [{CQCL.name: "name", CQCL.description: Services.care_home_service_with_nursing}], [{CQCL.name: "name", CQCL.code: "A1", CQCL.contacts: []}]),
+        ("loc_2", "prov_1", "non social care org", RegistrationStatus.registered, [{CQCL.name: "name", CQCL.description: Services.care_home_service_with_nursing}], [{CQCL.name: "name", CQCL.code: "A1", CQCL.contacts: []}]),
+        ("loc_3", "prov_1", None, RegistrationStatus.registered, [{CQCL.name: "name", CQCL.description: Services.care_home_service_with_nursing}], [{CQCL.name: "name", CQCL.code: "A1", CQCL.contacts: []}]),
+        ("loc_4", "prov_1", LocationType.social_care_identifier, RegistrationStatus.deregistered, [{CQCL.name: "name", CQCL.description: Services.care_home_service_with_nursing}], [{CQCL.name: "name", CQCL.code: "A1", CQCL.contacts: []}]),
+        ("loc_5", "prov_1", "non social care org", RegistrationStatus.deregistered, [{CQCL.name: "name", CQCL.description: Services.care_home_service_with_nursing}], [{CQCL.name: "name", CQCL.code: "A1", CQCL.contacts: []}]),
+        ("loc_6", "prov_1", None, RegistrationStatus.deregistered,  [{CQCL.name: "name", CQCL.description: Services.care_home_service_with_nursing}], [{CQCL.name: "name", CQCL.code: "A1", CQCL.contacts: []}]),
+        ("loc_7", "prov_1", LocationType.social_care_identifier, RegistrationStatus.registered, [{CQCL.name: "name", CQCL.description: Services.specialist_college_service}], [{CQCL.name: "name", CQCL.code: "A1", CQCL.contacts: []}]),
+        ("loc_8", "prov_1", "non social care org", RegistrationStatus.registered, [{CQCL.name: "name", CQCL.description: Services.specialist_college_service}], [{CQCL.name: "name", CQCL.code: "A1", CQCL.contacts: []}]),
+        ("loc_9", "prov_1", None, RegistrationStatus.registered, [{CQCL.name: "name", CQCL.description: Services.specialist_college_service}], [{CQCL.name: "name", CQCL.code: "A1", CQCL.contacts: []}]),
+        ("loc_10", "prov_1", LocationType.social_care_identifier, RegistrationStatus.deregistered, [{CQCL.name: "name", CQCL.description: Services.specialist_college_service}], [{CQCL.name: "name", CQCL.code: "A1", CQCL.contacts: []}]),
+        ("loc_11", "prov_1", "non social care org", RegistrationStatus.deregistered, [{CQCL.name: "name", CQCL.description: Services.specialist_college_service}], [{CQCL.name: "name", CQCL.code: "A1", CQCL.contacts: []}]),
+        ("loc_12", "prov_1", None, RegistrationStatus.deregistered,  [{CQCL.name: "name", CQCL.description: Services.specialist_college_service}], [{CQCL.name: "name", CQCL.code: "A1", CQCL.contacts: []}]),
+        ("loc_13", "prov_1", LocationType.social_care_identifier, RegistrationStatus.registered, [{CQCL.name: "name", CQCL.description: Services.care_home_service_with_nursing}], None),
+        ("loc_14", None, LocationType.social_care_identifier, RegistrationStatus.registered, [{CQCL.name: "name", CQCL.description: Services.care_home_service_with_nursing}], [{CQCL.name: "name", CQCL.code: "A1", CQCL.contacts: []}]),
+        (RecordsToRemoveInLocationsData.dental_practice, "prov_1", LocationType.social_care_identifier, RegistrationStatus.registered, None, [{CQCL.name: "name", CQCL.code: "A1", CQCL.contacts: []}]),
+        (RecordsToRemoveInLocationsData.temp_registration, "prov_1", LocationType.social_care_identifier, RegistrationStatus.registered, None, [{CQCL.name: "name", CQCL.code: "A1", CQCL.contacts: []}]),
+    ]
+    # fmt: on
+
+    identify_if_location_has_a_known_value_when_array_type_rows = [
+        ("loc_1", []),
+        ("loc_1", [{CQCL.name: "name", CQCL.code: "A1"}]),
+        ("loc_1", None),
+        ("loc_2", []),
+        ("loc_3", None),
+    ]
+    expected_identify_if_location_has_a_known_value_when_array_type_rows = [
+        ("loc_1", [], True),
+        ("loc_1", [{CQCL.name: "name", CQCL.code: "A1"}], True),
+        ("loc_1", None, True),
+        ("loc_2", [], False),
+        ("loc_3", None, False),
+    ]
+
+    identify_if_location_has_a_known_value_when_not_array_type_rows = [
+        ("loc_1", None),
+        ("loc_1", "prov_1"),
+        ("loc_1", None),
+        ("loc_2", None),
+    ]
+    expected_identify_if_location_has_a_known_value_when_not_array_type_rows = [
+        ("loc_1", None, True),
+        ("loc_1", "prov_1", True),
+        ("loc_1", None, True),
+        ("loc_2", None, False),
+    ]
+
+
+@dataclass
+class ValidateProvidersAPICleanedData:
+    raw_cqc_providers_rows = [
+        ("1-000000001", "20240101"),
+        ("1-000000002", "20240101"),
+        ("1-000000001", "20240201"),
+        ("1-000000002", "20240201"),
+    ]
+    cleaned_cqc_providers_rows = [
+        ("1-000000001", date(2024, 1, 1), "name", Sector.independent),
+        ("1-000000002", date(2024, 1, 1), "name", Sector.independent),
+        ("1-000000001", date(2024, 1, 9), "name", Sector.independent),
+        ("1-000000002", date(2024, 1, 9), "name", Sector.independent),
+    ]
+
+    calculate_expected_size_rows = raw_cqc_providers_rows
