+++ tests/test_file_schemas.py
@@ -131,786 +131,6 @@ class CapacityTrackerNonResSchema:
 
 
 @dataclass
-class CQCLocationsSchema:
-    detailed_schema = StructType(
-        [
-            StructField(CQCL.location_id, StringType(), True),
-            StructField(CQCL.provider_id, StringType(), True),
-            StructField(CQCL.organisation_type, StringType(), True),
-            StructField(CQCL.type, StringType(), True),
-            StructField(CQCL.name, StringType(), True),
-            StructField(CQCL.registration_status, StringType(), True),
-            StructField(CQCL.registration_date, StringType(), True),
-            StructField(CQCL.deregistration_date, StringType(), True),
-            StructField(CQCL.dormancy, StringType(), True),
-            StructField(CQCL.number_of_beds, IntegerType(), True),
-            StructField(CQCL.website, StringType(), True),
-            StructField(CQCL.postal_address_line1, StringType(), True),
-            StructField(CQCL.postal_address_town_city, StringType(), True),
-            StructField(CQCL.postal_address_county, StringType(), True),
-            StructField(CQCL.region, StringType(), True),
-            StructField(CQCL.postal_code, StringType(), True),
-            StructField(CQCL.onspd_latitude, StringType(), True),
-            StructField(CQCL.onspd_longitude, StringType(), True),
-            StructField(CQCL.care_home, StringType(), True),
-            StructField(CQCL.inspection_directorate, StringType(), True),
-            StructField(CQCL.main_phone_number, StringType(), True),
-            StructField(CQCL.local_authority, StringType(), True),
-            StructField(
-                CQCL.regulated_activities,
-                ArrayType(
-                    StructType(
-                        [
-                            StructField(CQCL.name, StringType(), True),
-                            StructField(CQCL.code, StringType(), True),
-                            StructField(
-                                CQCL.contacts,
-                                ArrayType(
-                                    StructType(
-                                        [
-                                            StructField(
-                                                CQCL.person_family_name,
-                                                StringType(),
-                                                True,
-                                            ),
-                                            StructField(
-                                                CQCL.person_given_name,
-                                                StringType(),
-                                                True,
-                                            ),
-                                            StructField(
-                                                CQCL.person_roles,
-                                                ArrayType(StringType(), True),
-                                                True,
-                                            ),
-                                            StructField(
-                                                CQCL.person_title, StringType(), True
-                                            ),
-                                        ]
-                                    )
-                                ),
-                                True,
-                            ),
-                        ]
-                    )
-                ),
-            ),
-            StructField(
-                CQCL.gac_service_types,
-                ArrayType(
-                    StructType(
-                        [
-                            StructField(CQCL.name, StringType(), True),
-                            StructField(CQCL.description, StringType(), True),
-                        ]
-                    )
-                ),
-            ),
-            StructField(
-                CQCL.specialisms,
-                ArrayType(
-                    StructType(
-                        [
-                            StructField(CQCL.name, StringType(), True),
-                        ]
-                    )
-                ),
-                True,
-            ),
-            StructField(
-                CQCL.current_ratings,
-                StructType(
-                    [
-                        StructField(
-                            CQCL.overall,
-                            StructType(
-                                [
-                                    StructField(
-                                        CQCL.organisation_id, StringType(), True
-                                    ),
-                                    StructField(CQCL.rating, StringType(), True),
-                                    StructField(CQCL.report_date, StringType(), True),
-                                    StructField(
-                                        CQCL.report_link_id, StringType(), True
-                                    ),
-                                    StructField(
-                                        CQCL.use_of_resources,
-                                        StructType(
-                                            [
-                                                StructField(
-                                                    CQCL.organisation_id,
-                                                    StringType(),
-                                                    True,
-                                                ),
-                                                StructField(
-                                                    CQCL.summary, StringType(), True
-                                                ),
-                                                StructField(
-                                                    CQCL.use_of_resources_rating,
-                                                    StringType(),
-                                                    True,
-                                                ),
-                                                StructField(
-                                                    CQCL.combined_quality_summary,
-                                                    StringType(),
-                                                    True,
-                                                ),
-                                                StructField(
-                                                    CQCL.combined_quality_rating,
-                                                    StringType(),
-                                                    True,
-                                                ),
-                                                StructField(
-                                                    CQCL.report_date,
-                                                    StringType(),
-                                                    True,
-                                                ),
-                                                StructField(
-                                                    CQCL.report_link_id,
-                                                    StringType(),
-                                                    True,
-                                                ),
-                                            ]
-                                        ),
-                                        True,
-                                    ),
-                                    StructField(
-                                        CQCL.key_question_ratings,
-                                        ArrayType(
-                                            StructType(
-                                                [
-                                                    StructField(
-                                                        CQCL.name, StringType(), True
-                                                    ),
-                                                    StructField(
-                                                        CQCL.rating,
-                                                        StringType(),
-                                                        True,
-                                                    ),
-                                                    StructField(
-                                                        CQCL.report_date,
-                                                        StringType(),
-                                                        True,
-                                                    ),
-                                                    StructField(
-                                                        CQCL.organisation_id,
-                                                        StringType(),
-                                                        True,
-                                                    ),
-                                                    StructField(
-                                                        CQCL.report_link_id,
-                                                        StringType(),
-                                                        True,
-                                                    ),
-                                                ]
-                                            ),
-                                            True,
-                                        ),
-                                        True,
-                                    ),
-                                ]
-                            ),
-                            True,
-                        ),
-                        StructField(
-                            CQCL.service_ratings,
-                            ArrayType(
-                                StructType(
-                                    [
-                                        StructField(CQCL.name, StringType(), True),
-                                        StructField(CQCL.rating, StringType(), True),
-                                        StructField(
-                                            CQCL.report_date, StringType(), True
-                                        ),
-                                        StructField(
-                                            CQCL.organisation_id, StringType(), True
-                                        ),
-                                        StructField(
-                                            CQCL.report_link_id, StringType(), True
-                                        ),
-                                        StructField(
-                                            CQCL.key_question_ratings,
-                                            ArrayType(
-                                                StructType(
-                                                    [
-                                                        StructField(
-                                                            CQCL.name,
-                                                            StringType(),
-                                                            True,
-                                                        ),
-                                                        StructField(
-                                                            CQCL.rating,
-                                                            StringType(),
-                                                            True,
-                                                        ),
-                                                    ]
-                                                ),
-                                                True,
-                                            ),
-                                            True,
-                                        ),
-                                    ]
-                                ),
-                                True,
-                            ),
-                            True,
-                        ),
-                    ]
-                ),
-                True,
-            ),
-            StructField(
-                CQCL.historic_ratings,
-                ArrayType(
-                    StructType(
-                        [
-                            StructField(CQCL.report_date, StringType(), True),
-                            StructField(CQCL.report_link_id, StringType(), True),
-                            StructField(CQCL.organisation_id, StringType(), True),
-                            StructField(
-                                CQCL.service_ratings,
-                                ArrayType(
-                                    StructType(
-                                        [
-                                            StructField(CQCL.name, StringType(), True),
-                                            StructField(
-                                                CQCL.rating, StringType(), True
-                                            ),
-                                            StructField(
-                                                CQCL.key_question_ratings,
-                                                ArrayType(
-                                                    StructType(
-                                                        [
-                                                            StructField(
-                                                                CQCL.name,
-                                                                StringType(),
-                                                                True,
-                                                            ),
-                                                            StructField(
-                                                                CQCL.rating,
-                                                                StringType(),
-                                                                True,
-                                                            ),
-                                                        ]
-                                                    ),
-                                                    True,
-                                                ),
-                                                True,
-                                            ),
-                                        ]
-                                    ),
-                                    True,
-                                ),
-                                True,
-                            ),
-                            StructField(
-                                CQCL.overall,
-                                StructType(
-                                    [
-                                        StructField(CQCL.rating, StringType(), True),
-                                        StructField(
-                                            CQCL.use_of_resources,
-                                            StructType(
-                                                [
-                                                    StructField(
-                                                        CQCL.combined_quality_rating,
-                                                        StringType(),
-                                                        True,
-                                                    ),
-                                                    StructField(
-                                                        CQCL.combined_quality_summary,
-                                                        StringType(),
-                                                        True,
-                                                    ),
-                                                    StructField(
-                                                        CQCL.use_of_resources_rating,
-                                                        StringType(),
-                                                        True,
-                                                    ),
-                                                    StructField(
-                                                        CQCL.use_of_resources_summary,
-                                                        StringType(),
-                                                        True,
-                                                    ),
-                                                ]
-                                            ),
-                                            True,
-                                        ),
-                                        StructField(
-                                            CQCL.key_question_ratings,
-                                            ArrayType(
-                                                StructType(
-                                                    [
-                                                        StructField(
-                                                            CQCL.name,
-                                                            StringType(),
-                                                            True,
-                                                        ),
-                                                        StructField(
-                                                            CQCL.rating,
-                                                            StringType(),
-                                                            True,
-                                                        ),
-                                                    ]
-                                                ),
-                                                True,
-                                            ),
-                                            True,
-                                        ),
-                                    ]
-                                ),
-                                True,
-                            ),
-                        ]
-                    ),
-                    True,
-                ),
-                True,
-            ),
-            StructField(
-                CQCL.relationships,
-                ArrayType(
-                    StructType(
-                        [
-                            StructField(CQCL.related_location_id, StringType(), True),
-                            StructField(CQCL.related_location_name, StringType(), True),
-                            StructField(CQCL.type, StringType(), True),
-                            StructField(CQCL.reason, StringType(), True),
-                        ]
-                    ),
-                    True,
-                ),
-                True,
-            ),
-            StructField(Keys.import_date, StringType(), True),
-        ]
-    )
-
-    impute_historic_relationships_schema = StructType(
-        [
-            StructField(CQCL.location_id, StringType(), True),
-            StructField(CQCLClean.cqc_location_import_date, DateType(), True),
-            StructField(CQCL.registration_status, StringType(), True),
-            StructField(
-                CQCL.relationships,
-                ArrayType(
-                    StructType(
-                        [
-                            StructField(CQCL.related_location_id, StringType(), True),
-                            StructField(CQCL.related_location_name, StringType(), True),
-                            StructField(CQCL.type, StringType(), True),
-                            StructField(CQCL.reason, StringType(), True),
-                        ]
-                    ),
-                    True,
-                ),
-                True,
-            ),
-        ]
-    )
-    expected_impute_historic_relationships_schema = StructType(
-        [
-            *impute_historic_relationships_schema,
-            StructField(
-                CQCLClean.imputed_relationships,
-                ArrayType(
-                    StructType(
-                        [
-                            StructField(CQCL.related_location_id, StringType(), True),
-                            StructField(CQCL.related_location_name, StringType(), True),
-                            StructField(CQCL.type, StringType(), True),
-                            StructField(CQCL.reason, StringType(), True),
-                        ]
-                    ),
-                    True,
-                ),
-                True,
-            ),
-        ]
-    )
-
-    get_relationships_where_type_is_predecessor_schema = StructType(
-        [
-            StructField(CQCL.location_id, StringType(), True),
-            StructField(CQCLClean.cqc_location_import_date, DateType(), True),
-            StructField(CQCL.registration_status, StringType(), True),
-            StructField(
-                CQCLClean.first_known_relationships,
-                ArrayType(
-                    StructType(
-                        [
-                            StructField(CQCL.related_location_id, StringType(), True),
-                            StructField(CQCL.related_location_name, StringType(), True),
-                            StructField(CQCL.type, StringType(), True),
-                            StructField(CQCL.reason, StringType(), True),
-                        ]
-                    ),
-                    True,
-                ),
-                True,
-            ),
-        ]
-    )
-    expected_get_relationships_where_type_is_predecessor_schema = StructType(
-        [
-            *get_relationships_where_type_is_predecessor_schema,
-            StructField(
-                CQCLClean.relationships_predecessors_only,
-                ArrayType(
-                    StructType(
-                        [
-                            StructField(CQCL.related_location_id, StringType(), True),
-                            StructField(CQCL.related_location_name, StringType(), True),
-                            StructField(CQCL.type, StringType(), True),
-                            StructField(CQCL.reason, StringType(), True),
-                        ]
-                    ),
-                    True,
-                ),
-                True,
-            ),
-        ]
-    )
-
-    impute_missing_struct_column_schema = StructType(
-        [
-            StructField(CQCL.location_id, StringType(), True),
-            StructField(CQCLClean.cqc_location_import_date, DateType(), True),
-            StructField(
-                CQCL.gac_service_types,
-                ArrayType(
-                    StructType(
-                        [
-                            StructField(CQCL.name, StringType(), True),
-                            StructField(CQCL.description, StringType(), True),
-                        ]
-                    )
-                ),
-            ),
-        ]
-    )
-    expected_impute_missing_struct_column_schema = StructType(
-        [
-            *impute_missing_struct_column_schema,
-            StructField(
-                CQCLClean.imputed_gac_service_types,
-                ArrayType(
-                    StructType(
-                        [
-                            StructField(CQCL.name, StringType(), True),
-                            StructField(CQCL.description, StringType(), True),
-                        ]
-                    )
-                ),
-            ),
-        ]
-    )
-
-    remove_locations_that_never_had_regulated_activities_schema = StructType(
-        [
-            StructField(CQCL.location_id, StringType(), True),
-            StructField(
-                CQCLClean.imputed_regulated_activities,
-                ArrayType(
-                    StructType(
-                        [
-                            StructField(CQCL.name, StringType(), True),
-                            StructField(CQCL.code, StringType(), True),
-                            StructField(
-                                CQCL.contacts,
-                                ArrayType(
-                                    StructType(
-                                        [
-                                            StructField(
-                                                CQCL.person_family_name,
-                                                StringType(),
-                                                True,
-                                            ),
-                                            StructField(
-                                                CQCL.person_given_name,
-                                                StringType(),
-                                                True,
-                                            ),
-                                            StructField(
-                                                CQCL.person_roles,
-                                                ArrayType(StringType(), True),
-                                                True,
-                                            ),
-                                            StructField(
-                                                CQCL.person_title, StringType(), True
-                                            ),
-                                        ]
-                                    )
-                                ),
-                                True,
-                            ),
-                        ]
-                    )
-                ),
-            ),
-        ]
-    )
-
-    extract_from_struct_schema = StructType(
-        [
-            StructField(CQCL.location_id, StringType(), True),
-            StructField(
-                CQCLClean.gac_service_types,
-                ArrayType(
-                    StructType(
-                        [
-                            StructField(CQCL.name, StringType(), True),
-                            StructField(CQCL.description, StringType(), True),
-                        ]
-                    )
-                ),
-            ),
-        ]
-    )
-    expected_extract_from_struct_schema = StructType(
-        [
-            *extract_from_struct_schema,
-            StructField(
-                CQCLClean.services_offered,
-                ArrayType(
-                    StringType(),
-                ),
-            ),
-        ]
-    )
-
-    primary_service_type_schema = StructType(
-        [
-            StructField(CQCL.location_id, StringType(), True),
-            StructField(CQCL.provider_id, StringType(), True),
-            StructField(
-                CQCLClean.imputed_gac_service_types,
-                ArrayType(
-                    StructType(
-                        [
-                            StructField(CQCL.name, StringType(), True),
-                            StructField(CQCL.description, StringType(), True),
-                        ]
-                    )
-                ),
-            ),
-        ]
-    )
-    expected_primary_service_type_schema = StructType(
-        [
-            StructField(CQCL.location_id, StringType(), True),
-            StructField(CQCL.provider_id, StringType(), True),
-            StructField(
-                CQCLClean.imputed_gac_service_types,
-                ArrayType(
-                    StructType(
-                        [
-                            StructField(CQCL.name, StringType(), True),
-                            StructField(CQCL.description, StringType(), True),
-                        ]
-                    )
-                ),
-            ),
-            StructField(CQCLClean.primary_service_type, StringType(), True),
-        ]
-    )
-
-    realign_carehome_column_schema = StructType(
-        [
-            StructField(CQCL.location_id, StringType(), True),
-            StructField(CQCL.care_home, StringType(), True),
-            StructField(CQCLClean.primary_service_type, StringType(), True),
-        ]
-    )
-
-    small_location_schema = StructType(
-        [
-            StructField(CQCL.location_id, StringType(), True),
-            StructField(CQCL.provider_id, StringType(), True),
-            StructField(Keys.import_date, StringType(), True),
-        ]
-    )
-
-    join_provider_schema = StructType(
-        [
-            StructField(CQCPClean.provider_id, StringType(), True),
-            StructField(CQCPClean.name, StringType(), True),
-            StructField(CQCPClean.cqc_sector, StringType(), True),
-            StructField(CQCPClean.cqc_provider_import_date, DateType(), True),
-        ]
-    )
-
-    expected_joined_schema = StructType(
-        [
-            StructField(CQCL.location_id, StringType(), True),
-            StructField(CQCL.provider_id, StringType(), True),
-            StructField(CQCLClean.provider_name, StringType(), True),
-            StructField(CQCPClean.cqc_sector, StringType(), True),
-            StructField(CQCLClean.cqc_location_import_date, DateType(), True),
-            StructField(CQCPClean.cqc_provider_import_date, DateType(), True),
-        ]
-    )
-
-    invalid_postcode_schema = StructType(
-        [
-            StructField(CQCL.location_id, StringType(), True),
-            StructField(CQCL.postal_code, StringType(), True),
-        ]
-    )
-
-    registration_status_schema = StructType(
-        [
-            StructField(CQCL.location_id, StringType(), True),
-            StructField(CQCL.registration_status, StringType(), True),
-        ]
-    )
-
-    social_care_org_schema = StructType(
-        [
-            StructField(CQCL.location_id, StringType(), True),
-            StructField(CQCL.type, StringType(), True),
-        ]
-    )
-
-    locations_for_ons_join_schema = StructType(
-        [
-            StructField(CQCLClean.location_id, StringType(), True),
-            StructField(CQCLClean.provider_id, StringType(), True),
-            StructField(CQCLClean.cqc_location_import_date, DateType(), True),
-            StructField(CQCLClean.postal_code, StringType(), True),
-            StructField(CQCL.registration_status, StringType(), True),
-        ]
-    )
-
-    ons_postcode_directory_schema = StructType(
-        [
-            StructField(ONSClean.postcode, StringType(), True),
-            StructField(ONSClean.contemporary_ons_import_date, DateType(), True),
-            StructField(ONSClean.contemporary_cssr, StringType(), True),
-            StructField(ONSClean.contemporary_region, StringType(), True),
-            StructField(ONSClean.current_ons_import_date, DateType(), True),
-            StructField(ONSClean.current_cssr, StringType(), True),
-            StructField(ONSClean.current_region, StringType(), True),
-        ]
-    )
-
-    expected_ons_join_schema = StructType(
-        [
-            StructField(ONSClean.contemporary_ons_import_date, DateType(), True),
-            StructField(CQCL.postal_code, StringType(), True),
-            StructField(CQCLClean.cqc_location_import_date, DateType(), True),
-            StructField(CQCL.location_id, StringType(), True),
-            StructField(CQCL.provider_id, StringType(), True),
-            StructField(ONSClean.contemporary_cssr, StringType(), True),
-            StructField(ONSClean.contemporary_region, StringType(), True),
-            StructField(ONSClean.current_ons_import_date, DateType(), True),
-            StructField(ONSClean.current_cssr, StringType(), True),
-            StructField(ONSClean.current_region, StringType(), True),
-            StructField(CQCL.registration_status, StringType(), True),
-        ]
-    )
-
-    expected_split_registered_schema = StructType(
-        [
-            *expected_ons_join_schema,
-        ]
-    )
-
-    clean_registration_column_schema = StructType(
-        [
-            StructField(CQCL.location_id, StringType(), True),
-            StructField(CQCL.registration_date, StringType(), True),
-            StructField(Keys.import_date, StringType(), True),
-        ]
-    )
-
-    expected_clean_registration_column_schema = StructType(
-        [
-            StructField(CQCL.location_id, StringType(), True),
-            StructField(CQCL.registration_date, StringType(), True),
-            StructField(Keys.import_date, StringType(), True),
-            StructField(CQCLClean.imputed_registration_date, StringType(), True),
-        ]
-    )
-
-    calculate_time_registered_for_schema = StructType(
-        [
-            StructField(CQCLClean.location_id, StringType(), True),
-            StructField(CQCLClean.cqc_location_import_date, DateType(), True),
-            StructField(CQCLClean.imputed_registration_date, DateType(), True),
-        ]
-    )
-    expected_calculate_time_registered_for_schema = StructType(
-        [
-            *calculate_time_registered_for_schema,
-            StructField(CQCLClean.time_registered, IntegerType(), True),
-        ]
-    )
-
-    remove_late_registration_dates_schema = StructType(
-        [
-            StructField(CQCL.location_id, StringType(), True),
-            StructField(Keys.import_date, StringType(), True),
-            StructField(CQCLClean.imputed_registration_date, StringType(), True),
-        ]
-    )
-    clean_provider_id_column_schema = StructType(
-        [
-            StructField(CQCL.location_id, StringType(), True),
-            StructField(CQCL.provider_id, StringType(), True),
-            StructField(Keys.import_date, StringType(), True),
-        ]
-    )
-
-    impute_missing_data_from_provider_dataset_schema = StructType(
-        [
-            StructField(CQCL.provider_id, StringType(), True),
-            StructField(CQCLClean.cqc_sector, StringType(), True),
-            StructField(CQCLClean.cqc_location_import_date, DateType(), True),
-        ]
-    )
-
-    remove_specialist_colleges_schema = StructType(
-        [
-            StructField(CQCL.location_id, StringType(), True),
-            StructField(
-                CQCLClean.services_offered,
-                ArrayType(
-                    StringType(),
-                ),
-            ),
-        ]
-    )
-    add_related_location_column_schema = StructType(
-        [
-            StructField(CQCL.location_id, StringType(), True),
-            StructField(
-                CQCLClean.imputed_relationships,
-                ArrayType(
-                    StructType(
-                        [
-                            StructField(CQCL.related_location_id, StringType(), True),
-                            StructField(CQCL.related_location_name, StringType(), True),
-                            StructField(CQCL.type, StringType(), True),
-                            StructField(CQCL.reason, StringType(), True),
-                        ]
-                    ),
-                    True,
-                ),
-                True,
-            ),
-        ]
-    )
-
-    expected_add_related_location_column_schema = StructType(
-        [
-            *add_related_location_column_schema,
-            StructField(CQCLClean.related_location, StringType(), True),
-        ]
-    )
-
-
-@dataclass
 class ExtractRegisteredManagerNamesSchema:
     extract_registered_manager_schema = StructType(
         [
@@ -1316,56 +536,6 @@ class CleaningUtilsSchemas:
         [
             *create_banded_bed_count_column_schema,
             StructField(IndCQC.number_of_beds_banded, FloatType(), True),
-        ]
-    )
-
-
-@dataclass
-class CQCProviderSchema:
-    rows_without_cqc_sector_schema = StructType(
-        [
-            StructField(CQCP.provider_id, StringType(), True),
-            StructField("some_data", StringType(), True),
-        ]
-    )
-    expected_rows_with_cqc_sector_schema = StructType(
-        [
-            StructField(CQCP.provider_id, StringType(), True),
-            StructField("some_data", StringType(), True),
-            StructField(CQCPClean.cqc_sector, StringType(), True),
-        ]
-    )
-
-    full_schema = StructType(
-        fields=[
-            StructField(CQCP.provider_id, StringType(), True),
-            StructField(
-                CQCP.location_ids,
-                ArrayType(
-                    StringType(),
-                ),
-            ),
-            StructField(CQCP.organisation_type, StringType(), True),
-            StructField(CQCP.ownership_type, StringType(), True),
-            StructField(CQCP.type, StringType(), True),
-            StructField(CQCP.uprn, StringType(), True),
-            StructField(CQCP.name, StringType(), True),
-            StructField(CQCP.registration_status, StringType(), True),
-            StructField(CQCP.registration_date, StringType(), True),
-            StructField(CQCP.deregistration_date, StringType(), True),
-            StructField(CQCP.postal_address_line1, StringType(), True),
-            StructField(CQCP.postal_address_town_city, StringType(), True),
-            StructField(CQCP.postal_address_county, StringType(), True),
-            StructField(CQCP.region, StringType(), True),
-            StructField(CQCP.postal_code, StringType(), True),
-            StructField(CQCP.onspd_latitude, FloatType(), True),
-            StructField(CQCP.onspd_longitude, FloatType(), True),
-            StructField(CQCP.main_phone_number, StringType(), True),
-            StructField(CQCP.companies_house_number, StringType(), True),
-            StructField(CQCP.inspection_directorate, StringType(), True),
-            StructField(CQCP.constituency, StringType(), True),
-            StructField(CQCP.local_authority, StringType(), True),
-            StructField(Keys.import_date, StringType(), True),
         ]
     )
 
@@ -3845,248 +3015,6 @@ class ValidationUtils:
             StructField(IndCQC.primary_service_type, StringType(), True),
         ]
     )
-
-
-@dataclass
-class ValidateLocationsAPICleanedData:
-    raw_cqc_locations_schema = StructType(
-        [
-            StructField(CQCL.location_id, StringType(), True),
-            StructField(CQCL.provider_id, StringType(), True),
-            StructField(Keys.import_date, StringType(), True),
-            StructField(CQCL.type, StringType(), True),
-            StructField(CQCL.registration_status, StringType(), True),
-            StructField(
-                CQCL.gac_service_types,
-                ArrayType(
-                    StructType(
-                        [
-                            StructField(CQCL.name, StringType(), True),
-                            StructField(CQCL.description, StringType(), True),
-                        ]
-                    )
-                ),
-            ),
-            StructField(
-                CQCL.regulated_activities,
-                ArrayType(
-                    StructType(
-                        [
-                            StructField(CQCL.name, StringType(), True),
-                            StructField(CQCL.code, StringType(), True),
-                            StructField(
-                                CQCL.contacts,
-                                ArrayType(
-                                    StructType(
-                                        [
-                                            StructField(
-                                                CQCL.person_family_name,
-                                                StringType(),
-                                                True,
-                                            ),
-                                            StructField(
-                                                CQCL.person_given_name,
-                                                StringType(),
-                                                True,
-                                            ),
-                                            StructField(
-                                                CQCL.person_roles,
-                                                ArrayType(StringType(), True),
-                                                True,
-                                            ),
-                                            StructField(
-                                                CQCL.person_title, StringType(), True
-                                            ),
-                                        ]
-                                    )
-                                ),
-                                True,
-                            ),
-                        ]
-                    )
-                ),
-            ),
-        ]
-    )
-    cleaned_cqc_locations_schema = StructType(
-        [
-            StructField(CQCLClean.location_id, StringType(), True),
-            StructField(CQCLClean.cqc_location_import_date, DateType(), True),
-            StructField(CQCLClean.cqc_provider_import_date, DateType(), True),
-            StructField(CQCLClean.care_home, StringType(), True),
-            StructField(CQCLClean.name, StringType(), True),
-            StructField(CQCLClean.provider_id, StringType(), True),
-            StructField(CQCLClean.provider_name, StringType(), True),
-            StructField(CQCLClean.cqc_sector, StringType(), True),
-            StructField(CQCLClean.registration_status, StringType(), True),
-            StructField(CQCLClean.imputed_registration_date, DateType(), True),
-            StructField(CQCLClean.dormancy, StringType(), True),
-            StructField(CQCLClean.number_of_beds, IntegerType(), True),
-            StructField(CQCLClean.primary_service_type, StringType(), True),
-            StructField(CQCLClean.contemporary_ons_import_date, DateType(), True),
-            StructField(CQCLClean.contemporary_cssr, StringType(), True),
-            StructField(CQCLClean.contemporary_region, StringType(), True),
-            StructField(CQCLClean.current_ons_import_date, DateType(), True),
-            StructField(CQCLClean.current_cssr, StringType(), True),
-            StructField(CQCLClean.current_region, StringType(), True),
-            StructField(CQCLClean.current_rural_urban_ind_11, StringType(), True),
-            StructField(CQCLClean.services_offered, ArrayType(StringType())),
-            StructField(
-                CQCLClean.imputed_regulated_activities,
-                ArrayType(
-                    StructType(
-                        [
-                            StructField(CQCL.name, StringType(), True),
-                            StructField(CQCL.code, StringType(), True),
-                            StructField(
-                                CQCL.contacts,
-                                ArrayType(
-                                    StructType(
-                                        [
-                                            StructField(
-                                                CQCL.person_family_name,
-                                                StringType(),
-                                                True,
-                                            ),
-                                            StructField(
-                                                CQCL.person_given_name,
-                                                StringType(),
-                                                True,
-                                            ),
-                                            StructField(
-                                                CQCL.person_roles,
-                                                ArrayType(StringType(), True),
-                                                True,
-                                            ),
-                                            StructField(
-                                                CQCL.person_title, StringType(), True
-                                            ),
-                                        ]
-                                    )
-                                ),
-                                True,
-                            ),
-                        ]
-                    )
-                ),
-            ),
-        ]
-    )
-    calculate_expected_size_schema = StructType(
-        [
-            StructField(CQCL.location_id, StringType(), True),
-            StructField(CQCL.provider_id, StringType(), True),
-            StructField(CQCL.type, StringType(), True),
-            StructField(CQCL.registration_status, StringType(), True),
-            StructField(
-                CQCL.gac_service_types,
-                ArrayType(
-                    StructType(
-                        [
-                            StructField(CQCL.name, StringType(), True),
-                            StructField(CQCL.description, StringType(), True),
-                        ]
-                    )
-                ),
-            ),
-            StructField(
-                CQCL.regulated_activities,
-                ArrayType(
-                    StructType(
-                        [
-                            StructField(CQCL.name, StringType(), True),
-                            StructField(CQCL.code, StringType(), True),
-                            StructField(
-                                CQCL.contacts,
-                                ArrayType(
-                                    StructType(
-                                        [
-                                            StructField(
-                                                CQCL.person_family_name,
-                                                StringType(),
-                                                True,
-                                            ),
-                                            StructField(
-                                                CQCL.person_given_name,
-                                                StringType(),
-                                                True,
-                                            ),
-                                            StructField(
-                                                CQCL.person_roles,
-                                                ArrayType(StringType(), True),
-                                                True,
-                                            ),
-                                            StructField(
-                                                CQCL.person_title, StringType(), True
-                                            ),
-                                        ]
-                                    )
-                                ),
-                                True,
-                            ),
-                        ]
-                    )
-                ),
-            ),
-        ]
-    )
-
-    identify_if_location_has_a_known_value_when_array_type_schema = StructType(
-        [
-            StructField(CQCL.location_id, StringType(), True),
-            StructField(
-                CQCL.regulated_activities,
-                ArrayType(
-                    StructType(
-                        [
-                            StructField(CQCL.name, StringType(), True),
-                            StructField(CQCL.code, StringType(), True),
-                        ]
-                    )
-                ),
-            ),
-        ]
-    )
-    expected_identify_if_location_has_a_known_value_when_array_type_schema = StructType(
-        [
-            *identify_if_location_has_a_known_value_when_array_type_schema,
-            StructField("has_known_value", BooleanType(), True),
-        ]
-    )
-
-    identify_if_location_has_a_known_value_when_not_array_type_schema = StructType(
-        [
-            StructField(CQCL.location_id, StringType(), True),
-            StructField(CQCL.provider_id, StringType(), True),
-        ]
-    )
-    expected_identify_if_location_has_a_known_value_when_not_array_type_schema = (
-        StructType(
-            [
-                *identify_if_location_has_a_known_value_when_not_array_type_schema,
-                StructField("has_known_value", BooleanType(), True),
-            ]
-        )
-    )
-
-
-@dataclass
-class ValidateProvidersAPICleanedData:
-    raw_cqc_providers_schema = StructType(
-        [
-            StructField(CQCP.provider_id, StringType(), True),
-            StructField(Keys.import_date, StringType(), True),
-        ]
-    )
-    cleaned_cqc_providers_schema = StructType(
-        [
-            StructField(CQCPClean.provider_id, StringType(), True),
-            StructField(CQCPClean.cqc_provider_import_date, DateType(), True),
-            StructField(CQCPClean.name, StringType(), True),
-            StructField(CQCPClean.cqc_sector, StringType(), True),
-        ]
-    )
-    calculate_expected_size_schema = raw_cqc_providers_schema
 
 
 @dataclass
